% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actions.R
\name{Actions}
\alias{Actions}
\alias{Actions.Region}
\title{Actions class builder}
\usage{
Actions(
  region,
  population_model,
  type = c("detection", "control", "removal"),
  stages = NULL,
  schedule = "all",
  class = character(),
  ...
)

\method{Actions}{Region}(
  region,
  population_model,
  type = c("detection", "control", "removal"),
  stages = NULL,
  schedule = "all",
  class = character(),
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the population spread model simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the population spread model
simulations.}

\item{type}{One of \code{"detection"} (default), \code{"control"},
or \code{"removal"} to indicate the type of actions applied.}

\item{stages}{Numeric vector of population stages (indices) to which
actions are applied. Default is all stages (when set to \code{NULL}).}

\item{schedule}{Vector of discrete simulation time steps (t = 0, 1, 2, ...)
in which to apply actions. Default is all time steps (when set to
\code{"all"}).}

\item{...}{Additional parameters.}
}
\value{
A \code{Actions} class object (list) containing a function for
applying simulated actions:
\describe{
\item{\code{get_type()}}{Get the actions type.}
\item{\code{get_id()}}{Get the actions numeric identifier.}
\item{\code{set_id(id)}}{Set the actions numeric identifier. Should be
an integer >= 1).}
\item{\code{get_label(include_id = TRUE)}}{Get the actions label used
in simulation results. Set \code{include_id} to include the action
\code{id} as a label prefix (default is \code{TRUE}).}
\item{\code{get_stages()}}{Get the population stages to which actions
are applied.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which actions are applied.}
\item{\code{include_cost()}}{Logical indication of a cost parameter
having a value.}
\item{\code{get_cost_label(include_id = TRUE)}}{Get the action cost
label used in simulation results. Set \code{include_id} to include
the action \code{id} as a label prefix (default is \code{TRUE}).}
\item{\code{get_cost_unit()}}{Get the unit of action cost.}
\item{\code{get_attributes(n)}}{Get attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return a list of the attached attributes.}
\item{\code{clear_attributes(n)}}{Clear attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return \code{n} without the attached attributes.}
\item{\code{apply(n, tm)}}{Apply actions to a simulated population
vector or matrix \code{n}, potentially with attached attributes
relating to previously applied actions, providing the time step
\code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied actions.}
}
}
\description{
Builds a generic class for simulating the application of invasive species
management actions, such as detection, control, and removal.
}
\references{
Baker, C. M., Bower, S., Tartaglia, E., Bode, M., Bower, H., & Pressey,
R. L. (2018). Modelling the spread and control of cherry guava on Lord
Howe Island. \emph{Biological Conservation}, 227, 252–258.
\doi{10.1016/j.biocon.2018.09.017}

Bradhurst, R., Spring, D., Stanaway, M., Milner, J., & Kompas, T. (2021).
A generalised and scalable framework for modelling incursions,
surveillance and control of plant and environmental pests.
\emph{Environmental Modelling & Software}, 139, N.PAG.
\doi{10.1016/j.envsoft.2021.105004}

Cacho, O. J., & Hester, S. M. (2022). Modelling biocontrol of invasive
insects: An application to European Wasp (Vespula germanica) in Australia.
\emph{Ecological Modelling}, 467. \doi{10.1016/j.ecolmodel.2022.109939}

Gormley, A. M., Holland, E. P., Barron, M. C., Anderson, D. P., & Nugent,
G. (2016). A modelling framework for predicting the optimal balance
between control and surveillance effort in the local eradication of
tuberculosis in New Zealand wildlife.
\emph{Preventive Veterinary Medicine}, 125, 10–18.
\doi{10.1016/j.prevetmed.2016.01.007}

Krug, R. M., Roura-Pascual, N., & Richardson, D. M. (2010). Clearing of
invasive alien plants under different budget scenarios: using a
simulation model to test efficiency. \emph{Biological Invasions}, 12(12),
4099–4112. \doi{10.1007/s10530-010-9827-3}

Lustig, A., James, A., Anderson, D., & Plank, M. (2019). Pest control at a
regional scale: Identifying key criteria using a spatially explicit,
agent‐based model. \emph{Journal of Applied Ecology}, 56(7 pp.1515–1527),
1527–1515. \doi{10.1111/1365-2664.13387}

Ramsey, D. S. L., & Efford, M. G. (2010). Management of bovine
tuberculosis in brushtail possums in New Zealand: predictions from a
spatially explicit individual-based model.
\emph{Journal of Applied Ecology}, 47(4), 911–919.
\doi{10.1111/j.1365-2664.2010.01839.x}

Rout, T. M., Moore, J. L., & McCarthy, M. A. (2014). Prevent, search or
destroy? A partially observable model for invasive species management.
\emph{Journal of Applied Ecology}, 51(3), 804–813.
\doi{10.1111/1365-2664.12234}

Spring, D., Croft, L., & Kompas, T. (2017). Look before you treat:
increasing the cost effectiveness of eradication programs with aerial
surveillance. \emph{Biological Invasions}, 19(2), 521.
\doi{10.1007/s10530-016-1292-1}

Wadsworth, R. A., Collingham, Y. C., Willis, S. G., Huntley, B., & Hulme,
P. E. (2000). Simulating the Spread and Management of Alien Riparian
Weeds: Are They Out of Control? \emph{Journal of Applied Ecology}, 37,
28–38. \doi{10.1046/j.1365-2664.2000.00551.x}

Warburton, B., & Gormley, A. M. (2015). Optimising the Application of
Multiple-Capture Traps for Invasive Species Management Using Spatial
Simulation. \emph{PLoS ONE}, 10(3), 1–14.
\doi{10.1371/journal.pone.0120373}

Zub, K., García-Díaz, P., Sankey, S., Eisler, R., & Lambin, X. (2022).
Using a Modeling Approach to Inform Progress Towards Stoat Eradication
From the Orkney Islands. \emph{Frontiers in Conservation Science}, 2.
\doi{10.3389/fcosc.2021.780102}
}
