% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controls.R
\name{Controls}
\alias{Controls}
\alias{Controls.Region}
\title{Controls class builder}
\usage{
Controls(
  region,
  population_model,
  control_type = c("search_destroy", "growth", "spread", "establishment"),
  manage_pr = NULL,
  manage_pr_type = c("individual", "population"),
  control_mult = NULL,
  exist_control = NULL,
  control_cost = NULL,
  radius = NULL,
  stages = NULL,
  apply_to = NULL,
  schedule = "all",
  ...
)

\method{Controls}{Region}(
  region,
  population_model,
  control_type = c("search_destroy", "growth", "spread", "establishment"),
  manage_pr = NULL,
  manage_pr_type = c("individual", "population"),
  control_mult = NULL,
  exist_control = NULL,
  control_cost = NULL,
  radius = NULL,
  stages = NULL,
  apply_to = NULL,
  schedule = "all",
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the population spread model simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the population spread model
simulations.}

\item{control_type}{One of \code{"search_destroy"} (default),
\code{"growth"} (including reproduction), \code{"spread"}, or
\code{"establishment"} to indicate the type of control applied.}

\item{manage_pr}{A vector of probability of management control success (or
effectiveness) values at each spatial location for combined surveillance
and removal approaches ("search & destroy", traps, etc.). Required when
\code{"search_destroy"} control type is selected. Default is \code{NULL}
for other control types.}

\item{manage_pr_type}{Configures how the management control effectiveness
(probability of control success) values are specified. May be specified
at either: the code{"individual"} level (default), whereby effectiveness
values denote the probability of control for each (controllable) invasive
individual; the \code{"population"} level, whereby effectiveness values
denote the probability of controlling all (controllable) individuals
within the (local) invasive population. Only utilised when
\code{"search_destroy"} control type is selected.}

\item{control_mult}{Control multipliers or rates required for control types
\code{"growth"}, \code{"spread"}, or \code{"establishment"}. May be a
single value (0-1), or vector of values for each location specified by
the \code{region}. Control multipliers are applied to reduce (with values
< 1) or suppress (with values = 0) growth, spread, or establishment at
locations (and optionally surroundings) where the invasive species has
been detected (when specified via a population attribute), as well as at
existing, known, or scheduled treatment locations when
\code{"exist_control"} is specified.}

\item{exist_control}{A vector to specify the distribution (logical mask) of
existing, known, or scheduled treatment locations for the reduction or
suppression of growth, spread, or establishment (spatially-explicit only).
Default is \code{NULL}. Only utilised when \code{"growth"},
\code{"spread"}, or \code{"establishment"} control type is selected.}

\item{control_cost}{Numeric vector of distributed control costs (combined
resource and fixed costs) or a single cost value for each location where
control is applied. For spatially-implicit area-based regions, cost
should be specified as either cost for the entire spatially-implicit area
for control type \code{"search_destroy"}, or cost per metres squared for
control type \code{"growth"}, \code{"spread"}, or \code{"establishment"}.
Costs are accumulated for each application of the control at each
(scheduled) simulation time step. The cost unit may be added as an
attribute (\code{attr(control_cost, "unit")}). Default is \code{NULL}
when costs are unavailable.}

\item{radius}{Optional radius (m) of the applied control for types
\code{"growth"}, \code{"spread"}, or \code{"establishment"} only.
Control is applied to all surrounding locations within the specified
radius of each location where the invasive species has been detected.
Default is \code{NULL}, indicating that control is only applied at
existing, known, or scheduled treatment locations, and/or at detected
locations (when specified via a population attribute).}

\item{stages}{Numeric vector of population stages (indices) to which
search & destroy or growth controls are applied. Default is all stages
(when set to \code{NULL}).}

\item{apply_to}{Optional label for growth control to indicate that it should
be only be applied to reproduction or survival rates. If applicable, set
to either \code{"reproduction"} or \code{"survival"}. If not specified,
control is applied to both reproduction and survival. The \code{stages}
and \code{apply_to} parameters may be utilised together, for example, to
control the seasonal survival rates of specified life stages.}

\item{schedule}{Vector of discrete simulation time steps (t = 0, 1, 2, ...)
in which to apply controls. Default is all time steps (when set to
\code{"all"}).}

\item{...}{Additional parameters.}
}
\value{
A \code{Controls} class object (list) containing a function
for accessing attributes and applying simulated controls:
\describe{
\item{\code{get_type()}}{Get the type of action ("control").}
\item{\code{get_id()}}{Get the actions numeric identifier.}
\item{\code{set_id(id)}}{Set the actions numeric identifier. Should be
an integer >= 1).}
\item{\code{get_label(include_id = TRUE)}}{Get the actions label used
in simulation results (e.g."\if{html}{\out{<id>}}_control_growth" or "control_growth").
Set \code{include_id} to include the action \code{id} as a label
prefix (default is \code{TRUE}).}
\item{\code{get_manage_pr_type()}}{Get the control effectiveness
(probability of management success) type.}
\item{\code{get_stages()}}{Get the population stages to which controls
are applied.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which controls are applied.}
\item{\code{include_cost()}}{Logical indication of a cost parameter
having a value.}
\item{\code{get_cost_label(include_id = TRUE)}}{Get the control
cost label used in simulation results ("\if{html}{\out{<id>}}_control_growth_cost" or
"control_growth_cost"). Set \code{include_id} to include the action
\code{id} as a label prefix (default is \code{TRUE}).}
\item{\code{get_cost_unit()}}{Get the unit of control cost.}
\item{\code{get_attributes(n)}}{Get attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return a list of the attached attributes.}
\item{\code{clear_attributes(n)}}{Clear attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return \code{n} without the attached attributes.}
\item{\code{apply(n, tm)}}{Apply controls to a simulated population
vector or matrix \code{n}, potentially with attached attributes
relating to previously applied actions, providing the time step
\code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied controls.}
}
}
\description{
Builds a class for simulating the application of invasive species management
controls, such as combined surveillance and removal ("search & destroy"),
and reducing or suppressing growth, reproduction, spread, or establishment.
}
\references{
Bradhurst, R., Spring, D., Stanaway, M., Milner, J., & Kompas, T. (2021).
A generalised and scalable framework for modelling incursions,
surveillance and control of plant and environmental pests.
\emph{Environmental Modelling & Software}, 139, N.PAG.
\doi{10.1016/j.envsoft.2021.105004}

Cacho, O. J., & Hester, S. M. (2022). Modelling biocontrol of invasive
insects: An application to European Wasp (Vespula germanica) in Australia.
\emph{Ecological Modelling}, 467. \doi{10.1016/j.ecolmodel.2022.109939}

Gormley, A. M., Holland, E. P., Barron, M. C., Anderson, D. P., & Nugent,
G. (2016). A modelling framework for predicting the optimal balance
between control and surveillance effort in the local eradication of
tuberculosis in New Zealand wildlife.
\emph{Preventive Veterinary Medicine}, 125, 10–18.
\doi{10.1016/j.prevetmed.2016.01.007}

Krug, R. M., Roura-Pascual, N., & Richardson, D. M. (2010). Clearing of
invasive alien plants under different budget scenarios: using a
simulation model to test efficiency. \emph{Biological Invasions}, 12(12),
4099–4112. \doi{10.1007/s10530-010-9827-3}

Lustig, A., James, A., Anderson, D., & Plank, M. (2019). Pest control at a
regional scale: Identifying key criteria using a spatially explicit,
agent‐based model. \emph{Journal of Applied Ecology}, 56(7 pp.1515–1527),
1527–1515. \doi{10.1111/1365-2664.13387}

Ramsey, D. S. L., & Efford, M. G. (2010). Management of bovine
tuberculosis in brushtail possums in New Zealand: predictions from a
spatially explicit individual-based model.
\emph{Journal of Applied Ecology}, 47(4), 911–919.
\doi{10.1111/j.1365-2664.2010.01839.x}

Rout, T. M., Moore, J. L., & McCarthy, M. A. (2014). Prevent, search or
destroy? A partially observable model for invasive species management.
\emph{Journal of Applied Ecology}, 51(3), 804–813.
\doi{10.1111/1365-2664.12234}

Spring, D., Croft, L., & Kompas, T. (2017). Look before you treat:
increasing the cost effectiveness of eradication programs with aerial
surveillance. \emph{Biological Invasions}, 19(2), 521.
\doi{10.1007/s10530-016-1292-1}

Wadsworth, R. A., Collingham, Y. C., Willis, S. G., Huntley, B., & Hulme,
P. E. (2000). Simulating the Spread and Management of Alien Riparian
Weeds: Are They Out of Control? \emph{Journal of Applied Ecology}, 37,
28–38. \doi{10.1046/j.1365-2664.2000.00551.x}

Warburton, B., & Gormley, A. M. (2015). Optimising the Application of
Multiple-Capture Traps for Invasive Species Management Using Spatial
Simulation. \emph{PLoS ONE}, 10(3), 1–14.
\doi{10.1371/journal.pone.0120373}

Zub, K., García-Díaz, P., Sankey, S., Eisler, R., & Lambin, X. (2022).
Using a Modeling Approach to Inform Progress Towards Stoat Eradication
From the Orkney Islands. \emph{Frontiers in Conservation Science}, 2.
\doi{10.3389/fcosc.2021.780102}
}
