% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffusion.R
\name{Diffusion}
\alias{Diffusion}
\title{Diffusion dispersal model class builder}
\usage{
Diffusion(
  region,
  population_model,
  dispersal_stages = NULL,
  diffusion_rate = NULL,
  diffusion_coeff = NULL,
  diffusion_threshold = 1,
  proportion = NULL,
  direction_function = NULL,
  attractors = list(),
  permeability = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for spread simulations. The region object
contains functionality for calculating path distances and directions,
permeability graphs, and structures to facilitate two-tier dispersal.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the spread simulations.}

\item{dispersal_stages}{Numeric vector of population stages (indices) that
disperse. Default is all stages (when set to \code{NULL}).}

\item{diffusion_rate}{The (asymptotic) speed of diffusion (in m per time
step). Default is \code{NULL} (resulting in no diffusion).}

\item{diffusion_coeff}{Optional coefficient of diffusion (in m^2 per time
step) for spatially implicit reaction-diffusion. Only used when
unstructured or stage-structured populations are defined within a single
patch region, and the asymptotic \code{diffusion_rate} is unknown (thus
not defined). Default is \code{NULL}.}

\item{diffusion_threshold}{Optional diffusion boundary threshold for
spatially implicit reaction-diffusion. Defined as a fraction (> 0 and
<= 1) of the initial population outside the diffusion boundary. Only used
when unstructured or stage-structured populations are defined within a
single patch region. Default is 1, resulting in diffusion limited to the
asymptotic diffusion rate, defined via the \code{diffusion_rate}
parameter, or calculated via the \code{diffusion_coeff} parameter.}

\item{proportion}{The proportion of the (unstructured or staged) population
that disperses at each time step, or the proportion of local presence-only
destinations selected for diffusive dispersal. Default is \code{NULL}
(producing no diffusive dispersal).}

\item{direction_function}{A function (or kernel) in the form
\code{function(directions)}, that calculates the (relative) probability of
dispersal for each direction (0-360 degrees) specified as an integer
vector. Default is none.}

\item{attractors}{List containing \code{Attractor} (or inherited) class
objects for spatially weighted dispersal, and/or containing a named
numeric multiplier \code{source_density = <numeric>}, which (when present)
is used to dynamically create an "attractor" based on the population
density (number/capacity) at each dispersal location source (origin) at
each simulation time step. Default is empty.}

\item{permeability}{A \code{Permeability} class (or inherited) class object
for representing spatial permeability or constraints. Default is none.}

\item{...}{Additional parameters.}
}
\value{
A \code{Diffusion} class object (list), containing inherited and
extended functions from the generic \code{Dispersal} class for accessing
attributes (of the function environment) and performing diffusive
dispersal:
\describe{
\item{\code{pack(n)}}{Packs a population vector or matrix \code{n} into
a list containing a vector of occupied \code{cells} (indices), the
\code{original} population values at the occupied locations only, the
\code{remaining} occupied values (initially a duplicate
\code{original}), and a vector or matrix for the \code{relocated}
population values at all locations (initially all zero).}
\item{\code{unpack(n)}}{Unpacks a population list by combining the
\code{remaining} and \code{relocated} population values to form a
new post-dispersal population vector or matrix.}
\item{\code{disperse(n)}}{Perform location dispersal on a list \code{n}
of vectors or matrices, representing the occupied \code{cells}
(indices), the \code{original} occupied populations, the
\code{remaining} occupied populations, and the \code{relocated}
populations (at all region cells), and return the transformed list of
vectors or matrices. The separation of original, remaining and
relocated populations enables multiple models for different dispersal
vectors to run in sequence. Spatially implicit diffusion attaches a
\code{diffusion_radius} attribute to \code{n} at each time step.}
\item{\code{get_diffusion_rate()}}{Get the defined or calculated
(asymptotic) speed of diffusion.}
}
}
\description{
Builds a class for representing diffusive dispersal for spread simulations
with functionality inherited from the generic dispersal model.
}
\details{
Diffusion simulates the local spread of populations into neighboring
locations (cells) a speed specified by the \code{diffusion_rate} parameter.
Diffusion may be simulated for presence-only, unstructured or stage-based
\code{populations}. For unstructured or stage-based populations, a specified
\code{proportion} of the population at each occupied (cell) location is
selected (sampled) for diffusive dispersal at each simulation time step.
Presence-only populations may diffuse into a specified \code{proportion} of
neighboring locations that are reachable at the specified diffusion rate
within each simulation time step. The probability of diffusive dispersal may
optionally be adjusted via \code{direction} functions and/or
\code{attractor} (layer) values, as well as \code{permeability} or
constraint layers, which are used to adjust (effective) distances between
cells and/or omit unreachable destinations. An optional establishment
likelihood (layer), which is configured via the population model, may be
applied to each diffusive dispersal, resulting in potential "deaths" of
individuals or unsuccessful presence-only dispersal events. The dispersal
functionality utilizes a wrapped population list of separate values for the
\code{original}, \code{remaining}, and \code{relocated} populations. This
separation enables multiple dispersal models, representing different
dispersal vectors, to be run in sequence. For example, local diffusion may
be combined with human-mediated and/or wind-based long-distant "jump"
dispersal. Spatially implicit diffusion is also provided via simple radial
diffusion for a presence-only single-patch population, or via
reaction-diffusion for unstructured or stage-based single patch populations.
}
\references{
Andow, D. A., Kareiva, P. M., Levin, S. A., & Okubo, A. (1990). Spread of
invading organisms. \emph{Landscape Ecology}, 4(2–3), 177.
\doi{10.1007/bf00132860}

Bradhurst, R., Spring, D., Stanaway, M., Milner, J., & Kompas, T. (2021).
A generalised and scalable framework for modelling incursions,
surveillance and control of plant and environmental pests.
\emph{Environmental Modelling & Software}, 139, N.PAG.
\doi{10.1016/j.envsoft.2021.105004}

Fisher, R. A. (1937). The wave of advance of advantageous genes.
\emph{Ann. Eugenics} 7, 355–369. \doi{10.1111/j.1469-1809.1937.tb02153.x}

Jongejans, E., Skarpaas, O., & Shea, K. (2008). Dispersal, demography and
spatial population models for conservation and control management.
\emph{Perspectives In Plant Ecology Evolution And Systematics}, 9(3–4),
153–170. \doi{10.1016/j.ppees.2007.09.005}

Robinet, C., Kehlenbeck, H., Kriticos, D. J., Baker, R. H. A.,
Battisti, A., Brunel, S., Dupin, M., Eyre, D., Faccoli, M., Ilieva, Z.,
Kenis, M., Knight, J., Reynaud, P., Yart, A., & van der Werf, W. (2012).
A Suite of Models to Support the Quantitative Assessment of Spread in Pest
Risk Analysis. \emph{PLoS ONE}, 7(10), 1–18.
\doi{10.1371/journal.pone.0043366}

Skellam, J. G. (1951). Random Dispersal in Theoretical Populations.
\emph{Biometrika}, 38(1/2), 196–218. \doi{10.2307/2332328}
}
