% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impacts.R
\name{Impacts}
\alias{Impacts}
\title{Impacts class builder}
\usage{
Impacts(
  region,
  population_model,
  impact_type = c("presence", "density", "area"),
  valuation_type = c("monetary", "non-monetary", "dynamic"),
  asset_name,
  asset_value,
  value_unit = NULL,
  loss_rate,
  discount_rate = NULL,
  stages = NULL,
  recovery_delay = NULL,
  dynamic_links = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the population spread model simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the population spread model
simulations.}

\item{impact_type}{One of \code{"presence"} (default), \code{"density"}, or
\code{"area"}, to indicate if the impact of a threat incursion is
calculated based on presence, density, or area of occupancy respectively.
Density-based impacts may only be used when the threat is modelled via
unstructured or staged-based populations with carrying capacity specified.
Area-based impacts are only applicable to spatially implicit models.}

\item{valuation_type}{One of \code{"monetary"} (default),
\code{"non-monetary"}, or \code{"dynamic"} to indicate the type of
valuation associated with the impacted asset or resource.}

\item{asset_name}{A descriptive name of the asset or resource impacted by
the threat incursion.}

\item{asset_value}{The periodic monetary value per simulation time step
(e.g. crop yield), or the non-monetary value (e.g. habitat condition),
of the asset or resource impacted by a threat incursion. This may be a
spatial layer (\code{terra::SpatRaster} or \code{raster::RasterLayer}),
or a vector of location values, for a grid or network-based model region.
For spatially implicit models, asset value should be expressed as value
per metres squared. Asset value for dynamic value impacts (e.g. plantation
biomass) represents the total value or asset quantity prior to threat
presence, which increasingly looses value when impacted by the threat,
and may be implicitly monetary (with ongoing not periodic value)
or non-monetary, as indicated by the \code{value_unit}.}

\item{value_unit}{A descriptive unit for asset value quantities. Default is
\code{NULL}, which is changed to "$" for monetary assets.}

\item{loss_rate}{A fractional loss of asset or resource value resulting from
a threat incursion or presence. For monetary value impacts the fractional
loss is applied to the periodic asset value at every time step that the
threat is present or until recovered. For non-monetary value impacts the
fractional loss to the asset value is maintained during threat presence or
until recovered. For dynamic value impacts the fractional loss is
repeatedly applied to the asset value or resource quantity, thus
increasingly reducing it, whilst the threat is present.}

\item{discount_rate}{An optional discount rate (per time interval) for
monetary asset values to estimate future values that account
for inflation. Typically the discounting uses market interest rates.
Discounted impact values are calculated by dividing the asset values by
\code{(1 + discount_rate)^time_int} for a specified time interval.
Default is \code{NULL}.}

\item{stages}{Numeric vector of population stages (indices) that
contribute to impacts when applicable. Default is all stages (when set to
\code{NULL}).}

\item{recovery_delay}{Number of simulation time steps that incursion impacts
continue to be in effect at previously occupied locations before the
asset value at the locations recover. Only available for spatially
explicit grid or network models. Default \code{NULL} assumes no delay.}

\item{dynamic_links}{Dynamic value impacts may optionally be linked to apply
(proportional) dynamic losses to threat suitability, carrying capacity,
and/or dispersal attraction when applicable, such as when a threat
utilises a resource or asset (e.g. as a food source). Configure dynamic
links via a vector of one or more strings from \code{"suitability"},
\code{"capacity"}, and/or \code{"attractors"}. Default \code{NULL}
assumes no dynamic links or that impact values are not dynamic.}

\item{...}{Additional parameters.}
}
\value{
A \code{Impacts} class object (list) containing functions for
performing impact calculations:
\describe{
\item{\code{get_id()}}{Get the impacts numeric identifier.}
\item{\code{set_id(id)}}{Set the impacts numeric identifier. Should be
an integer >= 1).}
\item{\code{get_impact_type()}}{Get the impact type.}
\item{\code{get_valuation_type()}}{Get the valuation type.}
\item{\code{get_asset_name()}}{Get the asset name.}
\item{\code{get_value_unit()}}{Get the asset value unit.}
\item{\code{get_attr_names(n)}}{Get the names of impact related
attributes attached to the population vector or matrix \code{n}.}
\item{\code{update_recovery_delay(n, tm)}}{Update the recovery delay
attribute attached to the population vector or matrix \code{n} at time
step \code{tm}, and return \code{n} with the updated recovery delay
attribute.}
\item{\code{calculate(n, tm)}}{Perform impact calculations resulting
from incursion population vector or matrix \code{n} at time step
\code{tm}, and return \code{n} with impact values attached.}
}
}
\description{
Builds a class for calculating incursion impacts during population spread
model simulations. Impact values may be monetary, non-monetary, or dynamic.
Monetary value impacts specify periodic assets (e.g. crop yields) that loose
value at each simulated time step (e.g. crop season) whilst a threat is
present or until recovered. Non-monetary value impacts specify ongoing
assets (e.g. habitat condition) that reduce in value by a fixed amount
whilst a threat is present or until recovered. Dynamic value impacts also
specify ongoing assets or resources (e.g. plantation biomass) that
increasingly loose value whilst a threat is present and optionally remain
impacted until recovered. Dynamic losses in asset value may optionally be
linked to apply (proportional) dynamic losses to threat suitability
(establishment likelihood), carrying capacity, and/or dispersal attraction
when applicable, such as when a threat utilises a resource or asset (e.g.
as a food source).
}
\references{
Arponen, A., Heikkinen, R. K., Thomas, C. D., & Moilanen, A. (2005). The
Value of Biodiversity in Reserve Selection: Representation, Species
Weighting, and Benefit Functions. \emph{Conservation Biology}, 19(6),
2009–2014. \doi{10.1111/j.1523-1739.2005.00218.x}

Costanza, R., D'Arge, R., De Groot, R., Farber, S., Grasso, M.,
Hannon, B., Limburg, K., Naeem, S., O'Neill, R. V., & Paruelo, J. (1998).
The value of ecosystem services: putting the issues in perspective.
\emph{Ecological Economics}, 25(E 1), 67–72.
\doi{10.1016/S0921-8009(98)00019-6}

Department of Agriculture, Fisheries and Forestry (DAFF) (2021). National
Environmental Biosecurity Response Agreement 2.0.
\url{https://www.agriculture.gov.au/biosecuritytrade/policy/emergency/nebra}

Dodd, A. J., Stoeckl, N., Baumgartner, J. B. & Kompas, T. F. (2020). Key
Result Summary: Valuing Australia's Biosecurity System. Tech. Rep. 170713,
\emph{Centre of Excellence for Biosecurity Risk Analysis (CEBRA),}
\emph{The University of Melbourne, Melbourne}.
\url{https://cebra.unimelb.edu.au/__data/assets/pdf_file/0020/3535013/CEBRA_Value_Docs_KeyResultSummary_v0.6_Endorsed.pdf}

MacLeod, A., Head, J., & Gaunt, A. (2004). An assessment of the potential
economic impact of Thrips palmi on horticulture in England and the
significance of a successful eradication campaign. \emph{Crop Protection},
23(7), 610–601. \doi{10.1016/j.cropro.2003.11.010}

O'Loughlin, L. S., Gooden, B., Barney, J. N., & Lindenmayer, D. B. (2019).
Surrogacy in invasion research and management : inferring "impact" from
"invasiveness". \emph{Frontiers in Ecology and the Environment}, 17(8),
464–473. \doi{10.1002/fee.2097}

Soliman, T., Mourits, M. C. M., Oude Lansink, A. G. J. M., &
van der Werf, W. (2010). Economic impact assessment in pest risk analysis.
\emph{Crop Protection}, 29(6), 517–524. \doi{10.1016/j.cropro.2009.12.014}

Soliman, T., Mourits, M. C. M., Oude Lansink, A. G. J. M., &
van der Werf, W. (2015). Quantitative economic impact assessment of
invasive plant pests: What does it require and when is it worth the
effort? \emph{Crop Protection}, 69, 9–17.
\doi{10.1016/j.cropro.2014.11.011}

Stoeckl, N., Dodd, A., & Kompas, T. (2023). The monetary value of 16
services protected by the Australian National Biosecurity System:
Spatially explicit estimates and vulnerability to incursions.
\emph{Ecosystem Services}, 60. \doi{10.1016/j.ecoser.2023.101509}

Weitzman, M. L. (1998). The Noah's Ark Problem. \emph{Econometrica},
66(6), 1279–1298. \doi{10.2307/2999617}

Welsh, M. J., Brockerhoff, E. G., Kean, J. M., Phillips, C.,
Stringer, L. D., Vereijssen, J., Turner, J. A., Epanchin-Niell, R. S.,
Monge, J. J., Soliman, T., Robinson, A. P., Kompas, T., Liebhold, A. M.,
& Ormsby, M. (2021). Approaches for estimating benefits and costs of
interventions in plant biosecurity across invasion phases.
\emph{Ecological Applications}, 31(5). \doi{10.1002/eap.2319}
}
