% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incursions.R
\name{Incursions}
\alias{Incursions}
\alias{Incursions.Raster}
\alias{Incursions.SpatRaster}
\title{Incursions class builder}
\usage{
Incursions(
  x,
  type = c("weight", "prob"),
  continued = FALSE,
  time_steps = 1,
  ...
)

\method{Incursions}{Raster}(x, ...)

\method{Incursions}{SpatRaster}(
  x,
  type = c("weight", "prob"),
  continued = FALSE,
  time_steps = 1,
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the incursion weightings or arrival probabilities at
each spatial location.}

\item{type}{One of \code{"weight"} or \code{"prob"} to indicate if the
values in \code{x} represent single incursion weightings or arrival
probabilities respectively.}

\item{continued}{Logical to indicate whether incursions based on arrival
probabilities (when \code{type = "prob"}) continue to be generated during
spread simulations.}

\item{time_steps}{The time interval in simulation steps for generating
subsequent incursions when \code{type = "prob"} and
\code{continued = TRUE}.}

\item{...}{Additional parameters.}
}
\value{
An \code{Incursions} class object (list) containing functions for
accessing attributes (of the function environment) TODO:
\describe{
\item{\code{TODO()}}{TODO.}
}
}
\description{
Builds a class to generate invasive species incursions, either by selecting
a single incursion location via stochastic sampling based on weightings for
each available spatial location, or by sampling potentially multiple
locations via incursion or arrival probabilities defined for each location,
which optionally may continue to generate incursions at specified time
intervals during a spread simulation.
}
