% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initializer.R
\name{Initializer}
\alias{Initializer}
\alias{Initializer.Raster}
\alias{Initializer.SpatRaster}
\alias{Initializer.Incursions}
\title{Initializer class builder}
\usage{
Initializer(
  x,
  region = NULL,
  population_model = NULL,
  class = character(),
  ...
)

\method{Initializer}{Raster}(x, ...)

\method{Initializer}{SpatRaster}(
  x,
  region = NULL,
  population_model = NULL,
  class = character(),
  ...
)

\method{Initializer}{Incursions}(
  x,
  region = NULL,
  population_model = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster} defining
the initial population distribution, or an \code{Incursions} or inherited
class object for generating initial, and optionally continued, invasive
species incursions.}

\item{region}{A \code{Region} or inherited class object defining the spatial
locations included in the spread simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation and growth functionality for the
spread simulations.}

\item{class}{Character class name for inherited classes. Default is empty.}

\item{...}{Additional parameters.}
}
\value{
An \code{Initializer} class object (list) containing functions for
accessing attributes (of the function environment) TODO:
\describe{
\item{\code{TODO()}}{TODO.}
}
}
\description{
Builds a class to initialize a spread simulation, given a specified spatial
region and population representation, via either a predefined initial
population distribution, or via stochastic generation of invasive species
incursions.
}
