% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StagedPopulation.R
\name{StagedPopulation}
\alias{StagedPopulation}
\title{Stage or age-based population model class builder}
\usage{
StagedPopulation(
  region,
  growth,
  capacity = NULL,
  capacity_area = NULL,
  capacity_stages = NULL,
  establish_pr = NULL,
  incursion_mean = NULL,
  incursion_stages = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the spread simulations.}

\item{growth}{A stage or age-based matrix containing reproduction and
survival rates for each time step. An optional attribute \code{survivals}
may be used to differentiate the survival rates, otherwise survival rates
are assumed to be any non-zero values on the lower (left) triangle of the
matrix, including those on the diagonal.}

\item{capacity}{A vector of carrying capacity values of the invasive species
at each location specified by the \code{region}, or per unit area defined
by \code{capacity_area} when the \code{region} is spatially implicit.
Default is \code{NULL} for when growth is not capacity-limited.}

\item{capacity_area}{Carrying capacity area (in m^2) specified for
capacity-limited growth in a spatially implicit (single patch)
\code{region}. For example, use a value of \code{1e+06} when
\code{capacity} is specified in km^2. Default is \code{NULL}, although a
value is required when \code{capacity} is specified for a spatially
implicit region.}

\item{capacity_stages}{A vector of stage or age indices (as per the
rows/columns of \code{growth} matrix) indicating which life-stages/ages
are applicable for capacity-limited growth (and survival). Default is
\code{NULL} for when growth is not capacity-limited, or when all
stages/ages are applicable for capacity-limited growth.}

\item{establish_pr}{An optional vector of probability values (0-1) to
represent the likelihood of establishment at each location specified by
the \code{region}. This may be used to avoid transient/unsuccessful
incursions or migrations from being presented in the simulation results.
Default is \code{NULL}.}

\item{incursion_mean}{Numeric mean population size for incursion locations.
The population size is sampled from the Poisson distribution for each
incursion location.}

\item{incursion_stages}{A vector of stage or age indices (as per the
rows/columns of \code{growth} matrix) indicating which life-stages/ages
are applicable for incursions. Default is \code{NULL} for when all
stages/ages are applicable for incursions.}

\item{...}{Additional parameters.}
}
\value{
A \code{StagedPopulation} class object (list) containing functions
for accessing attributes and simulating growth:
\describe{
\item{\code{get_region()}}{Get the region object.}
\item{\code{get_type()}}{Get the population representation type.}
\item{\code{get_growth()}}{Get the growth stage/age matrix.}
\item{\code{get_growth_r()}}{Get the equivalent (single value) growth
rate for the \code{growth} stage/age matrix.}
\item{\code{get_capacity()}}{Get the carrying capacity as a vector of
values for each location.}
\item{\code{get_capacity_stages()}}{Get the stage/age (indices) that are
applicable for capacity-limited growth.}
\item{\code{get_establish_pr()}}{Get the establishment probability as a
vector of values for each location.}
\item{\code{make(initial, current, incursion)}}{Make a population matrix
(with a row per location and a column per stage/age) via using vectors
or matrices of the \code{initial} or \code{current} and
\code{incursion} population at each region location.}
\item{\code{grow(x)}}{Performs logistic (capacity-limited) growth on the
population \code{x} matrix (having a row per location and a column per
stage/age), and returns the transformed matrix.}
}
}
\description{
Builds a class for representing stage or age-based populations in spread
simulations. Simulates logistic (capacity-limited) growth.
}
\references{
Beverton, R. J. H., & Holt, S. J. (1957). On the dynamics of exploited
fish populations. \emph{Fisheries Investigations}, 19, 1-533.

García Adeva, J. J., Botha, J. H., & Reynolds, M. (2012). A simulation
modelling approach to forecast establishment and spread of Bactrocera
fruit flies. \emph{Ecological Modelling}, 227, 93–108.
\doi{10.1016/j.ecolmodel.2011.11.026}

Jongejans, E., Skarpaas, O., & Shea, K. (2008). Dispersal, demography and
spatial population models for conservation and control management.
\emph{Perspectives In Plant Ecology Evolution And Systematics}, 9(3–4),
153–170. \doi{10.1016/j.ppees.2007.09.005}

Lefkovitch, L. P. (1965). The Study of Population Growth in Organisms
Grouped by Stages. \emph{Biometrics}, 21(1), 1–18.
\doi{doi.org/10.2307/2528348}

Ricker, W. E. (1958). Handbook of computations for biological statistics
of fish populations. \emph{Bulletin (Fisheries Research Board of Canada)},
119. \url{https://waves-vagues.dfo-mpo.gc.ca/Library/1485.pdf}
}
