% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actions.R
\name{Actions}
\alias{Actions}
\alias{Actions.Region}
\title{Actions class builder}
\usage{
Actions(
  region,
  population_model,
  type = c("detection", "control", "removal"),
  stages = NULL,
  schedule = NULL,
  class = character(),
  ...
)

\method{Actions}{Region}(
  region,
  population_model,
  type = c("detection", "control", "removal"),
  stages = NULL,
  schedule = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the population spread model simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the population spread model
simulations.}

\item{type}{One of \code{"detection"} (default), \code{"control"},
or \code{"removal"} to indicate the type of actions applied.}

\item{stages}{Numeric vector of population stages (indices) to which
actions are applied. Default is all stages (when set to \code{NULL}).}

\item{schedule}{Vector of discrete simulation time steps (t = 0, 1, 2, ...)
in which to apply actions. Default is all time steps (when set to
\code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{Actions} class object (list) containing a function for
applying simulated actions:
\describe{
\item{\code{get_type()}}{Get the actions type.}
\item{\code{get_id()}}{Get the actions numeric identifier.}
\item{\code{set_id(id)}}{Set the actions numeric identifier. Should be
an integer >= 1).}
\item{\code{get_label(include_id = TRUE)}}{Get the actions label used
in simulation results. Set \code{include_id} to include the action
\code{id} as a label prefix (default is \code{TRUE}).}
\item{\code{get_stages()}}{Get the population stages to which actions
are applied.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which actions are applied.}
\item{\code{include_cost()}}{Logical indication of a cost parameter
having a value.}
\item{\code{get_cost_label(include_id = TRUE)}}{Get the action cost
label used in simulation results. Set \code{include_id} to include
the action \code{id} as a label prefix (default is \code{TRUE}).}
\item{\code{get_cost_unit()}}{Get the unit of action cost.}
\item{\code{get_attributes(n)}}{Get attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return a list of the attached attributes.}
\item{\code{clear_attributes(n)}}{Clear attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return \code{n} without the attached attributes.}
\item{\code{apply(n, tm)}}{Apply actions to a simulated population
vector or matrix \code{n}, potentially with attached attributes
relating to previously applied actions, providing the time step
\code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied actions.}
}
}
\description{
Builds a generic class for simulating the application of invasive species
management actions, such as detection, control, and removal.
}
