% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attractor.R
\name{Attractor}
\alias{Attractor}
\alias{Attractor.Raster}
\alias{Attractor.SpatRaster}
\alias{Attractor.numeric}
\title{Attractor class builder}
\usage{
Attractor(
  x,
  region,
  type = c("source", "destination", "both"),
  multiplier = NULL,
  ...
)

\method{Attractor}{Raster}(x, ...)

\method{Attractor}{SpatRaster}(
  x,
  region,
  type = c("source", "destination", "both"),
  multiplier = NULL,
  ...
)

\method{Attractor}{numeric}(
  x,
  region,
  type = c("source", "destination", "both"),
  multiplier = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object, or numeric vector, representing spatial attractor values across
the simulation region.}

\item{region}{A \code{Region} or inherited class object defining the spatial
locations included in the spread simulations.}

\item{type}{One of \code{"source"}, \code{"destination"}, or \code{"both"},
indicating whether attractor values represent scale factors for dispersal
from source locations, or dispersal likelihood weightings for destination
locations, or both.}

\item{multiplier}{Optional numeric multiplier for scaling or weighting
attractor values appropriately (for convenience).}

\item{...}{Additional parameters.}
}
\value{
A \code{Attractor} class object (list) containing functions for
accessing attributes:
\describe{
\item{\code{get_rast()}}{Get the attractor \code{terra::SpatRaster}
object.}
\item{\code{get_values(cells = NULL, na.incl = FALSE)}}{Get the
attractor values (multiplied by the specified multiplier when present)
for optionally specified region cell indices, which should only refer
to all raster indices (including NA-value cells) when specified (via
\code{na.incl = TRUE}), otherwise the index sequence should skip
NA-value cells (as per \code{Region get_indices}). Default is all
non-NA cell values.}
\item{\code{get_aggr_values(cells = NULL, na.incl = FALSE)}}{Get the
raster aggregated attractor values (multiplied by the specified
multiplier when present) for optionally specified aggregated region
cell indices, which should only refer to all raster indices (including
NA-value cells) when specified (via \code{na.incl = TRUE}), otherwise
the index sequence should skip NA-value cells. Default is all
non-NA cell values. Only available for an attractor built with a
spatial raster layer, and functional when region is two-tiered.}
\item{\code{get_type())}}{Get the attractor type ("source",
"destination", or "both").}
}
}
\description{
Builds a class to represent spatially distributed weights or scaling factors
assigned to cells/locations to represent an "attractor" that modify the
likelihood of dispersal from source or to destination locations. Attractor
values may be used to scale the proportion of each population dispersing
from each source location, or to scale the number of dispersal events
originating at each source location. Alternatively, attractor weight values
may be used to modify the relative likelihood of dispersal to each
destination location. The \code{type} attribute of an attractor may thus be
defined as "source", "destination", or "both".
}
\references{
Bossenbroek, J. M., Kraft, C. E., & Nekola, J. C. (2001). Prediction of
Long-Distance Dispersal Using Gravity Models: Zebra Mussel Invasion of
Inland Lakes. \emph{Ecological Applications}, 11(6), 1778–1788.
\doi{10.2307/3061095}

Carrasco, L. R., Mumford, J. D., MacLeod, A., Harwood, T.,
Grabenweger, G., Leach, A. W., Knight, J. D., & Baker, R. H. A. (2010).
Unveiling human-assisted dispersal mechanisms in invasive alien insects:
Integration of spatial stochastic simulation and phenology models.
\emph{Ecological Modelling}, 221(17), 2068–2075.
\doi{10.1016/j.ecolmodel.2010.05.012}

Crespo-Pérez, V., Rebaudo, F., Silvain, J.-F., & Dangles, O. (2011).
Modeling invasive species spread in complex landscapes: the case of potato
moth in Ecuador. \emph{Landscape Ecology}, 26(10), 1447.
\doi{10.1007/s10980-011-9649-4}

Muirhead, J. R., Leung, B., Overdijk, C., Kelly, D. W., Nandakumar, K.,
Marchant, K. R., & MacIsaac, H. J. (2006). Modelling local and
long-distance dispersal of invasive emerald ash borer Agrilus planipennis
(Coleoptera) in North America. \emph{Diversity & Distributions}, 12(1),
71–79. \doi{10.1111/j.1366-9516.2006.00218.x}
}
