% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attractor.R
\name{Attractor}
\alias{Attractor}
\alias{Attractor.Raster}
\alias{Attractor.SpatRaster}
\title{Attractor class builder}
\usage{
Attractor(x, region, ...)

\method{Attractor}{Raster}(x, ...)

\method{Attractor}{SpatRaster}(x, region, ...)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing spatial attractor values across the simulation region.}

\item{region}{A \code{Region} or inherited class object defining the spatial
locations included in the spread simulations.}

\item{...}{Additional parameters.}
}
\value{
A \code{Attractor} class object (list) containing functions for
accessing attributes:
\describe{
\item{\code{get_rast()}}{Get the attractor \code{terra::SpatRaster}
object.}
\item{\code{get_values(cells = NULL)}}{Get the attractor values for
optionally specified region (non-NA) cell indices. Default is all.}
}
}
\description{
Builds a class to represent spatially distributed weights, or actual or
relative probabilities assigned to cells/locations to represent an
"attractor" that influences the likelihood of dispersal from or to other
locations.
}
