% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controls.R
\name{Controls}
\alias{Controls}
\alias{Controls.Region}
\title{Controls class builder}
\usage{
Controls(
  region,
  population_model,
  control_type = c("search_destroy", "growth", "spread", "establishment"),
  manage_pr = NULL,
  manage_pr_type = c("individual", "population"),
  control_mult = NULL,
  exist_control = NULL,
  control_cost = NULL,
  radius = NULL,
  stages = NULL,
  apply_to = NULL,
  schedule = NULL,
  ...
)

\method{Controls}{Region}(
  region,
  population_model,
  control_type = c("search_destroy", "growth", "spread", "establishment"),
  manage_pr = NULL,
  manage_pr_type = c("individual", "population"),
  control_mult = NULL,
  exist_control = NULL,
  control_cost = NULL,
  radius = NULL,
  stages = NULL,
  apply_to = NULL,
  schedule = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the population spread model simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the population spread model
simulations.}

\item{control_type}{One of \code{"search_destroy"} (default),
\code{"growth"} (including reproduction), \code{"spread"}, or
\code{"establishment"} to indicate the type of control applied.}

\item{manage_pr}{A vector of probability of management control success (or
effectiveness) values at each spatial location for combined surveillance
and removal approaches ("search & destroy", traps, etc.). Required when
\code{"search_destroy"} control type is selected. Default is \code{NULL}
for other control types.}

\item{manage_pr_type}{Configures how the management control effectiveness
(probability of control success) values are specified. May be specified
at either: the code{"individual"} level (default), whereby effectiveness
values denote the probability of control for each (controllable) invasive
individual; the \code{"population"} level, whereby effectiveness values
denote the probability of controlling all (controllable) individuals
within the (local) invasive population. Only utilised when
\code{"search_destroy"} control type is selected.}

\item{control_mult}{Control multipliers or rates required for control types
\code{"growth"}, \code{"spread"}, or \code{"establishment"}. May be a
single value (0-1), or vector of values for each location specified by
the \code{region}. Control multipliers are applied to reduce (with values
< 1) or suppress (with values = 0) growth, spread, or establishment at
locations (and optionally surroundings) where the invasive species has
been detected (when specified via a population attribute), as well as at
existing, known, or scheduled treatment locations when
\code{"exist_control"} is specified.}

\item{exist_control}{A vector to specify the distribution (logical mask) of
existing, known, or scheduled treatment locations for the reduction or
suppression of growth, spread, or establishment (spatially-explicit only).
Default is \code{NULL}. Only utilised when \code{"growth"},
\code{"spread"}, or \code{"establishment"} control type is selected.}

\item{control_cost}{Numeric vector of distributed control costs (combined
resource and fixed costs) or a single cost value for each location where
control is applied. For spatially-implicit area-based regions, cost
should be specified as either cost for the entire spatially-implicit area
for control type \code{"search_destroy"}, or cost per metres squared for
control type \code{"growth"}, \code{"spread"}, or \code{"establishment"}.
Costs are accumulated for each application of the control at each
(scheduled) simulation time step. The cost unit may be added as an
attribute (\code{attr(control_cost, "unit")}). Default is \code{NULL}
when costs are unavailable.}

\item{radius}{Optional radius (m) of the applied control for types
\code{"growth"}, \code{"spread"}, or \code{"establishment"} only.
Control is applied to all surrounding locations within the specified
radius of each location where the invasive species has been detected.
Default is \code{NULL}, indicating that control is only applied at
existing, known, or scheduled treatment locations, and/or at detected
locations (when specified via a population attribute).}

\item{stages}{Numeric vector of population stages (indices) to which
search & destroy or growth controls are applied. Default is all stages
(when set to \code{NULL}).}

\item{apply_to}{Optional label for growth control to indicate that it should
be only be applied to reproduction or survival rates. If applicable, set
to either \code{"reproduction"} or \code{"survival"}. If not specified,
control is applied to both reproduction and survival. The \code{stages}
and \code{apply_to} parameters may be utilised together, for example, to
control the seasonal survival rates of specified life stages.}

\item{schedule}{Vector of discrete simulation time steps (t = 0, 1, 2, ...)
in which to apply controls. Default is all time steps (when set to
\code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{Controls} class object (list) containing a function
for accessing attributes and applying simulated controls:
\describe{
\item{\code{get_type()}}{Get the type of action ("control").}
\item{\code{get_id()}}{Get the actions numeric identifier.}
\item{\code{set_id(id)}}{Set the actions numeric identifier. Should be
an integer >= 1).}
\item{\code{get_label(include_id = TRUE)}}{Get the actions label used
in simulation results (e.g."\if{html}{\out{<id>}}_control_growth" or "control_growth").
Set \code{include_id} to include the action \code{id} as a label
prefix (default is \code{TRUE}).}
\item{\code{get_manage_pr_type()}}{Get the control effectiveness
(probability of management success) type.}
\item{\code{get_stages()}}{Get the population stages to which controls
are applied.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which controls are applied.}
\item{\code{include_cost()}}{Logical indication of a cost parameter
having a value.}
\item{\code{get_cost_label(include_id = TRUE)}}{Get the control
cost label used in simulation results ("\if{html}{\out{<id>}}_control_growth_cost" or
"control_growth_cost"). Set \code{include_id} to include the action
\code{id} as a label prefix (default is \code{TRUE}).}
\item{\code{get_cost_unit()}}{Get the unit of control cost.}
\item{\code{clear_attributes(n)}}{Clear attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return \code{n} without the attached attributes.}
\item{\code{apply(n, tm)}}{Apply controls to a simulated population
vector or matrix \code{n}, potentially with attached attributes
relating to previously applied actions, providing the time step
\code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied controls.}
}
}
\description{
Builds a class for simulating the application of invasive species management
controls, such as combined surveillance and removal ("search & destroy"),
and reducing or suppressing growth, reproduction, spread, or establishment.
}
