% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detection.R
\name{Detection}
\alias{Detection}
\alias{Detection.Region}
\title{Detection class builder}
\usage{
Detection(
  region,
  population_model,
  sensitivity,
  sensitivity_type = c("individual", "population", "presence"),
  sensitivity_threshold = NULL,
  surv_cost = NULL,
  stages = NULL,
  schedule = NULL,
  ...
)

\method{Detection}{Region}(
  region,
  population_model,
  sensitivity,
  sensitivity_type = c("individual", "population", "presence"),
  sensitivity_threshold = NULL,
  surv_cost = NULL,
  stages = NULL,
  schedule = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the population spread model simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the population spread model
simulations.}

\item{sensitivity}{Numeric vector of distributed surveillance sensitivities,
or probability of detection at each location.}

\item{sensitivity_type}{Configures how the surveillance detection
sensitivity values are specified. May be specified at either: the
\code{"individual"} level (default), whereby sensitivity values denote the
probability of detection for each (detectable) invasive individual; the
\code{"population"} size (threshold) level, whereby sensitivity values
denote the probability of detecting (local) populations having at
least at certain size (of detectable individuals), specified via a
threshold value (below), and the probability of detection for (local)
population sizes below the threshold value is reduced proportionally (i.e.
scaled by size/threshold); the \code{"presence"}/absence level, whereby
sensitivity values denote the probability of detection any (local)
presence of (detectable individuals within) the invasive species
population (equivalent to threshold = 1).}

\item{sensitivity_threshold}{The threshold (minimum) population size (of
detectable individuals) for (locally) applying surveillance detection
probability (sensitivity) values when \code{sensitivity_type} is
\code{"population"}. The probability of detection for (local) population
sizes below the threshold value is reduced proportionally (i.e. scaled by
size/threshold). Default is \code{NULL} when \code{sensitivity_type} is
\code{"population"}, or is set to \code{1} when \code{sensitivity_type} is
\code{"presence"}.}

\item{surv_cost}{Numeric vector of distributed surveillance costs (combined
resource and fixed costs) or a single cost value for each location where
surveillance is applied. For spatially-implicit area-based regions, cost
should be specified as cost for the entire spatially-implicit area. Costs
are accumulated for each application of the surveillance at each
(scheduled) simulation time step. The cost unit may be added as an
attribute (\code{attr(surv_cost, "unit")}). Default is \code{NULL} when
costs are unavailable.}

\item{stages}{Numeric vector of population stages (indices) which the
surveillance detects. Default is all stages (when set to \code{NULL}).}

\item{schedule}{Vector of discrete simulation time steps (t = 0, 1, 2, ...)
in which to apply surveillance. Default is all time steps (when set to
\code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{Detection} class object (list) containing a function
for accessing attributes and applying simulated detection:
\describe{
\item{\code{get_type()}}{Get the type of action ("detection").}
\item{\code{get_id()}}{Get the actions numeric identifier.}
\item{\code{set_id(id)}}{Set the actions numeric identifier. Should be
an integer >= 1).}
\item{\code{get_label(include_id = TRUE)}}{Get the actions label used
in simulation results ("\if{html}{\out{<id>}}_detected" or "detected").
Set \code{include_id} to include the action \code{id} as a label
prefix (default is \code{TRUE}).}
\item{\code{get_sensitivity()}}{Get the surveillance sensitivity.}
\item{\code{get_sensitivity_type()}}{Get the surveillance sensitivity
type.}
\item{\code{get_stages()}}{Get the population stages which the
surveillance detects.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which surveillance is applied.}
\item{\code{include_cost()}}{Logical indication of a cost parameter
having a value.}
\item{\code{get_cost_label(include_id = TRUE)}}{Get the surveillance
cost label used in simulation results ("\if{html}{\out{<id>}}_surv_cost" or
"surv_cost"). Set \code{include_id} to include the action \code{id}
as a label prefix (default is \code{TRUE}).}
\item{\code{get_cost_unit()}}{Get the unit of surveillance cost.}
\item{\code{get_attributes(n)}}{Get attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return a list of the attached attributes.}
\item{\code{clear_attributes(n)}}{Clear attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return \code{n} without the attached attributes.}
\item{\code{apply(n, tm)}}{Apply surveillance detection to a simulated
population vector or matrix \code{n}, potentially with attached
attributes relating to previously applied actions, providing the time
step \code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied detection/surveillance.}
}
}
\description{
Builds a class for simulating the application of detection or surveillance
applied in the management of an invasive species.
}
\references{
Baker, C. M., Bower, S., Tartaglia, E., Bode, M., Bower, H., & Pressey,
R. L. (2018). Modelling the spread and control of cherry guava on Lord
Howe Island. \emph{Biological Conservation}, 227, 252–258.
\doi{10.1016/j.biocon.2018.09.017}

Bradhurst, R., Spring, D., Stanaway, M., Milner, J., & Kompas, T. (2021).
A generalised and scalable framework for modelling incursions,
surveillance and control of plant and environmental pests.
\emph{Environmental Modelling & Software}, 139, N.PAG.
\doi{10.1016/j.envsoft.2021.105004}

Cacho, O. J., & Hester, S. M. (2022). Modelling biocontrol of invasive
insects: An application to European Wasp (Vespula germanica) in Australia.
\emph{Ecological Modelling}, 467. \doi{10.1016/j.ecolmodel.2022.109939}

Gormley, A. M., Holland, E. P., Barron, M. C., Anderson, D. P., & Nugent,
G. (2016). A modelling framework for predicting the optimal balance
between control and surveillance effort in the local eradication of
tuberculosis in New Zealand wildlife.
\emph{Preventive Veterinary Medicine}, 125, 10–18.
\doi{10.1016/j.prevetmed.2016.01.007}

Lustig, A., James, A., Anderson, D., & Plank, M. (2019). Pest control at a
regional scale: Identifying key criteria using a spatially explicit,
agent‐based model. \emph{Journal of Applied Ecology}, 56(7 pp.1515–1527),
1527–1515. \doi{10.1111/1365-2664.13387}

Ramsey, D. S. L., & Efford, M. G. (2010). Management of bovine
tuberculosis in brushtail possums in New Zealand: predictions from a
spatially explicit individual-based model.
\emph{Journal of Applied Ecology}, 47(4), 911–919.
\doi{10.1111/j.1365-2664.2010.01839.x}

Spring, D., Croft, L., & Kompas, T. (2017). Look before you treat:
increasing the cost effectiveness of eradication programs with aerial
surveillance. \emph{Biological Invasions}, 19(2), 521.
\doi{10.1007/s10530-016-1292-1}

Wadsworth, R. A., Collingham, Y. C., Willis, S. G., Huntley, B., & Hulme,
P. E. (2000). Simulating the Spread and Management of Alien Riparian
Weeds: Are They Out of Control? \emph{Journal of Applied Ecology}, 37,
28–38. \doi{10.1046/j.1365-2664.2000.00551.x}
}
