% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detection.R
\name{Detection}
\alias{Detection}
\alias{Detection.Region}
\title{Detection class builder}
\usage{
Detection(
  region,
  population_model,
  sensitivity,
  sensitivity_type = c("individual", "population", "presence"),
  sensitivity_threshold = NULL,
  surv_cost = NULL,
  stages = NULL,
  schedule = NULL,
  ...
)

\method{Detection}{Region}(
  region,
  population_model,
  sensitivity,
  sensitivity_type = c("individual", "population", "presence"),
  sensitivity_threshold = NULL,
  surv_cost = NULL,
  stages = NULL,
  schedule = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the population spread model simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the population spread model
simulations.}

\item{sensitivity}{Numeric vector of distributed surveillance sensitivities,
or probability of detection at each location.}

\item{sensitivity_type}{Configures how the surveillance detection
sensitivity values are specified. May be specified at either: the
\code{"individual"} level (default), whereby sensitivity values denote the
probability of detection for each (detectable) invasive individual; the
\code{"population"} size (threshold) level, whereby sensitivity values
denote the probability of detecting (local) populations having at
least at certain size (of detectable individuals), specified via a
threshold value (below), and the probability of detection for (local)
population sizes below the threshold value is reduced proportionally (i.e.
scaled by size/threshold); the \code{"presence"}/absence level, whereby
sensitivity values denote the probability of detection any (local)
presence of (detectable individuals within) the invasive species
population (equivalent to threshold = 1).}

\item{sensitivity_threshold}{The threshold (minimum) population size (of
detectable individuals) for (locally) applying surveillance detection
probability (sensitivity) values when \code{sensitivity_type} is
\code{"population"}. The probability of detection for (local) population
sizes below the threshold value is reduced proportionally (i.e. scaled by
size/threshold). Default is \code{NULL} when \code{sensitivity_type} is
\code{"population"}, or is set to \code{1} when \code{sensitivity_type} is
\code{"presence"}.}

\item{surv_cost}{Numeric vector of distributed surveillance costs (combined
resource and fixed costs) or a single cost value for each location where
surveillance is applied. For spatially-implicit area-based regions, cost
should be specified as cost for the entire spatially-implicit area. Costs
are accumulated for each application of the surveillance at each
(scheduled) simulation time step. The cost unit may be added as an
attribute (\code{attr(surv_cost, "unit")}). Default is \code{NULL} when
costs are unavailable.}

\item{stages}{Numeric vector of population stages (indices) which the
surveillance detects. Default is all stages (when set to \code{NULL}).}

\item{schedule}{Vector of discrete simulation time steps (t = 0, 1, 2, ...)
in which to apply surveillance. Default is all time steps (when set to
\code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{Detection} class object (list) containing a function
for accessing attributes and applying simulated detection:
\describe{
\item{\code{get_type()}}{Get the type of action ("detection").}
\item{\code{get_id()}}{Get the actions numeric identifier.}
\item{\code{set_id(id)}}{Set the actions numeric identifier. Should be
an integer >= 1).}
\item{\code{get_label(include_id = TRUE)}}{Get the actions label used
in simulation results ("\if{html}{\out{<id>}}_detected" or "detected").
Set \code{include_id} to include the action \code{id} as a label
prefix (default is \code{TRUE}).}
\item{\code{get_sensitivity()}}{Get the surveillance sensitivity.}
\item{\code{get_sensitivity_type()}}{Get the surveillance sensitivity
type.}
\item{\code{get_stages()}}{Get the population stages which the
surveillance detects.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which surveillance is applied.}
\item{\code{include_cost()}}{Logical indication of a cost parameter
having a value.}
\item{\code{get_cost_label(include_id = TRUE)}}{Get the surveillance
cost label used in simulation results ("\if{html}{\out{<id>}}_surv_cost" or
"surv_cost"). Set \code{include_id} to include the action \code{id}
as a label prefix (default is \code{TRUE}).}
\item{\code{get_cost_unit()}}{Get the unit of surveillance cost.}
\item{\code{get_attributes(n)}}{Get attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return a list of the attached attributes.}
\item{\code{clear_attributes(n)}}{Clear attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return \code{n} without the attached attributes.}
\item{\code{apply(n, tm)}}{Apply surveillance detection to a simulated
population vector or matrix \code{n}, potentially with attached
attributes relating to previously applied actions, providing the time
step \code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied detection/surveillance.}
}
}
\description{
Builds a class for simulating the application of detection or surveillance
applied in the management of an invasive species.
}
