% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dispersal.R
\name{Dispersal}
\alias{Dispersal}
\alias{Dispersal.Region}
\title{Dispersal model abstract class builder}
\usage{
Dispersal(region, population_model, class = character(), ...)

\method{Dispersal}{Region}(region, class = character(), ...)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for spread simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the spread simulations.}

\item{class}{Character class name for inherited classes. Default is empty.}

\item{...}{Additional parameters.}
}
\value{
A \code{Dispersal} class object (list) containing functions for
accessing attributes (of the function environment) and performing
dispersal:
\describe{
\item{\code{get_region()}}{Get the spatial region object.}
\item{\code{get_population_model()}}{Get the population model object.}
\item{\code{disperse(n)}}{Perform location dispersal on a list \code{n}
of vectors or matrices, representing the occupied \code{cells}
(indices), the \code{original} occupied populations, the
\code{remaining} occupied populations, and the \code{relocated}
populations (at all region cells), and return the transformed list of
vectors or matrices. The separation of original, remaining and relocated
populations enables multiple models for different dispersal vectors to
run in sequence.}
\item{\code{TODO()}}{TODO.}
}
}
\description{
Builds a class for dispersal representation and functionality for spread
simulations.
}
