% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dispersal.R
\name{Dispersal}
\alias{Dispersal}
\alias{Dispersal.Raster}
\alias{Dispersal.SpatRaster}
\alias{Dispersal.Region}
\title{Dispersal model abstract class builder}
\usage{
Dispersal(region, ...)

\method{Dispersal}{Raster}(region, ...)

\method{Dispersal}{SpatRaster}(region, ...)

\method{Dispersal}{Region}(region, ...)
}
\arguments{
\item{region}{A \code{raster::RasterLayer}, \code{terra::SpatRaster}, or
\code{Region} or inherited class object representing the spatial region
(template) for spread simulations.}

\item{...}{Additional parameters.}
}
\value{
An \code{Dispersal} class object (list) containing functions for
accessing attributes (of the function environment) TODO:
\describe{
\item{\code{TODO()}}{TODO.}
}
}
\description{
Builds a class for dispersal representation and functionality for the
spread simulations.
}
