% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dispersal.R
\name{Dispersal}
\alias{Dispersal}
\alias{Dispersal.Region}
\title{Generic dispersal model class builder}
\usage{
Dispersal(
  region,
  population_model,
  dispersal_stages = NULL,
  proportion = NULL,
  events = NULL,
  distance_function = NULL,
  distance_adjust = NULL,
  direction_function = NULL,
  attractors = list(),
  permeability = NULL,
  max_distance = NULL,
  class = character(),
  ...
)

\method{Dispersal}{Region}(
  region,
  population_model,
  dispersal_stages = NULL,
  proportion = NULL,
  events = NULL,
  distance_function = NULL,
  distance_adjust = NULL,
  direction_function = NULL,
  attractors = list(),
  permeability = NULL,
  max_distance = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
grid-based spatial region (template) or patch-based network for spread
simulations. The region object contains functionality for calculating path
distances and directions, permeability graphs, and structures to
facilitate two-tier grid-based dispersal.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the spread simulations.}

\item{dispersal_stages}{Numeric vector of population stages (indices) that
disperse. Default is all stages (when set to \code{NULL}).}

\item{proportion}{The proportion of the (unstructured or staged) population
that disperses at each time step. This parameter is also used to scale the
the number of dispersal destinations selected when the population is
presence-only and the number of dispersal \code{events} is not defined.
Default is \code{NULL} (producing no dispersal unless the population is
presence-only and \code{events} is defined).}

\item{events}{The mean number of dispersal events generated via a Poisson
distribution for each location at each time step. A dispersal destination
(location) is selected for each dispersal event. Default is \code{NULL}
(resulting in destinations being selected for each individual within
unstructured or staged populations, or stochastic sampling of destinations
for presence-only populations).}

\item{distance_function}{A function (or kernel) in the form
\code{function(distances)}, that calculates the (relative) probability of
dispersal for each distance (in m) specified as a numeric vector. Default
is none.}

\item{distance_adjust}{Logical indication of whether the (relative)
probabilities returned by \code{distance_function} should be distributed
across the (approximate) number of grid cells at each distance. When not
specified (\code{NULL}), the value will resolve to \code{TRUE} for
grid-based regions only.}

\item{direction_function}{A function (or kernel) in the form
\code{function(directions)}, that calculates the (relative) probability of
dispersal for each direction (0-360 degrees) specified as an integer
vector. Default is none.}

\item{attractors}{List containing \code{Attractor} (or inherited) class
objects for spatially weighted dispersal, and/or containing a named
numeric multiplier \code{source_density = <numeric>}, which (when present)
is used to dynamically create an "attractor" based on the population
density (number/capacity) at each dispersal location source (origin) at
each simulation time step. Default is empty.}

\item{permeability}{A \code{Permeability} class (or inherited) class object
for representing spatial permeability or constraints. Default is none.}

\item{max_distance}{The maximum dispersal distance (in m). Default is
\code{NULL} (resulting in no explicit distance limit).}

\item{class}{Character class name for inherited classes. Default is
\code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{Dispersal} class object (list) containing functions for
accessing attributes (of the function environment) and performing
dispersal:
\describe{
\item{\code{set_cores(cores)}}{Set the number of cores available for
parallel processing and thus enable parallel processing for
calculating path distances and directions.}
\item{\code{pack(n)}}{Packs a population vector or matrix \code{n} into
a list containing a vector of occupied location (cell or patch)
\code{indices}, the \code{original} population values at the occupied
locations only, the \code{remaining} occupied values (initially a
duplicate of \code{original}), and a vector or matrix for the
\code{relocated} population values at all locations (initially all
zero).}
\item{\code{unpack(n)}}{Unpacks a population list by combining the
\code{remaining} and \code{relocated} population values to form a
new post-dispersal population vector or matrix.}
\item{\code{disperse(n)}}{Perform location dispersal on a list \code{n}
of vectors or matrices, representing the occupied location (cell or
patch) \code{indices}, the \code{original} occupied populations, the
\code{remaining} occupied populations, and the \code{relocated}
populations (at all region locations), and return the transformed list
of vectors or matrices. The separation of original, remaining and
relocated populations enables multiple models for different dispersal
vectors to run in sequence.}
}
}
\description{
Builds a generic class for representing dispersal with functionality for
spread simulations. Dispersal may be simulated for presence-only,
unstructured or stage-based \code{populations}. For unstructured or
stage-based populations, a specified \code{proportion} of the population at
each occupied location (cell or patch) is selected (sampled) for dispersal
at each simulation time step. Presence-only populations may disperse via
specifying a (mean) number of dispersal \code{events}. Dispersal events are
generated for each occupied location. Unstructured and stage-based population
dispersal may also utilize specified events, otherwise an event is assigned
to each dispersing individual. For each dispersal event a destination
location is selected via stochastic sampling using (relative) probabilities
of dispersal from each occupied location to reachable destinations. These
(relative) probabilities are calculated via \code{distance} and/or
\code{direction} functions, optionally combined with \code{attractor}
(layer) values. Paths to reachable destinations are calculated via the
\code{region} class object. Paths may be derived via \code{permeability} or
constraint grid-based layers or patch-based data, which are used to adjust
(effective) path distances and/or omit unreachable paths. Dispersal paths
may also be limited to a \code{maximum} (adjusted) distance. An optional
establishment likelihood (layer), which is configured via the population
model, may be applied to each dispersal event, resulting in potential
"deaths" of individuals or unsuccessful presence-only dispersal events.
Presence-only population dispersal may also be configured without
\code{events} when \code{proportion} is specified to represent a scaling
multiplier for (presumed) actual (rather than relative) dispersal
probabilities. Under these circumstances destination locations are sampled
from all reachable destinations using the scaled probabilities. The
dispersal functionality utilizes a wrapped population list of separate
values for the \code{original}, \code{remaining}, and \code{relocated}
populations. This separation enables multiple dispersal models, representing
different dispersal vectors, to be run in sequence.
}
