% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dispersal.R
\name{Dispersal}
\alias{Dispersal}
\alias{Dispersal.Region}
\title{Dispersal model abstract class builder}
\usage{
Dispersal(region, class = character(), ...)

\method{Dispersal}{Region}(region, class = character(), ...)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for spread simulations.}

\item{class}{Character class name for inherited classes. Default is empty.}

\item{...}{Additional parameters.}
}
\value{
A \code{Dispersal} class object (list) containing functions for
accessing attributes (of the function environment) and performing
dispersal:
\describe{
\item{\code{get_region()}}{Get the spatial region object.}
\item{\code{disperse(x)}}{Perform location dispersal on population
\code{x} array (stages by locations), and return transformed array.}
\item{\code{TODO()}}{TODO.}
}
}
\description{
Builds a class for dispersal representation and functionality for the
spread simulations.
}
