% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impacts.R
\name{Impacts}
\alias{Impacts}
\title{Impacts class builder}
\usage{
Impacts(
  impacts,
  population_model,
  impact_stages = NULL,
  calc_total = NULL,
  dynamic_links = NULL,
  recovery_delay = NULL,
  ...
)
}
\arguments{
\item{impacts}{A \code{bsimpact::ImpactAnalysis} or inherited class object
specifying how environment, social, and/or economic impacts are
calculated, classified, and/or combined based on asset values,
classifications, and/or incursion loss rates.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the population spread model
simulations.}

\item{impact_stages}{Numeric vector of population stages (indices) that
contribute to impacts. Default is all stages (when set to \code{NULL}).}

\item{calc_total}{Logical indicator for whether or not total impacts may
be (sensibly) calculated by summing across region locations. Default is
\code{NULL}, whereby totals are only calculated when the impact (context)
valuation type is \code{"monetary"}. Set to \code{TRUE} if it makes sense
to sum total \code{"non-monetary"} or \code{"ranking"} impact valuations
across region locations (providing multiple aspects can be combined).}

\item{dynamic_links}{Dynamic impacts (when \code{impacts} is a
\code{bsimpact::ValueImpacts} object with \code{is_dynamic = TRUE}) may
optionally be linked to apply (proportional) dynamic losses to threat
suitability, carrying capacity, and/or dispersal attraction when
applicable, such as when a threat utilises a resource or asset (e.g. as a
food source). Configure dynamic links via a vector of one or more strings
from \code{"suitability"}, \code{"capacity"}, and/or \code{"attractors"}.
Default \code{NULL} assumes no dynamic links or impacts are not dynamic.}

\item{recovery_delay}{Number of simulation time steps that incursion impacts
continue to be in effect at previously occupied locations before the
asset value at the locations recover. Only available for spatially
explicit grid or network models. Default \code{NULL} assumes no delay.}

\item{...}{Additional parameters.}
}
\value{
A \code{Impacts} class object (list) containing functions for
getting the impact context and performing impact calculations:
\describe{
\item{\code{get_impacts()}}{Get \code{bsimpact::ImpactAnalysis} or
inherited class object.}
\item{\code{get_context()}}{Get \code{bsimpact::Context} object.}
\item{\code{get_calc_total()}}{Get calculate total indicator.}
\item{\code{includes_combined()}}{Logical indicator for when impacts are
combined.}
\item{\code{update_recovery_delay(n)}}{Update the recovery delay
attribute attached to the population vector.}
\item{\code{calculate(n, tm)}}{Perform impact calculations resulting
from incursion population vector or matrix \code{n} at time step
\code{tm}, and return \code{n} with a list of impact values attached,
including combined impacts when applicable.}
}
}
\description{
Builds a class for calculating incursion impacts during population spread
model simulations. Encapsulates \code{bsimpact} impact analysis objects.
}
