% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incursions.R
\name{Incursions}
\alias{Incursions}
\alias{Incursions.Raster}
\alias{Incursions.SpatRaster}
\alias{Incursions.default}
\title{Incursions class builder}
\usage{
Incursions(
  x,
  region = NULL,
  type = c("weight", "prob"),
  continued = FALSE,
  time_steps = 1,
  ...
)

\method{Incursions}{Raster}(x, ...)

\method{Incursions}{SpatRaster}(x, region = NULL, ...)

\method{Incursions}{default}(
  x,
  region = NULL,
  type = c("weight", "prob"),
  continued = FALSE,
  time_steps = 1,
  ...
)
}
\arguments{
\item{x}{A \code{raster::RasterLayer}, \code{terra::SpatRaster}, or numeric
vector defining the incursion weightings or arrival probabilities at each
spatial location.}

\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for spread simulations.}

\item{type}{One of \code{"weight"} or \code{"prob"} to indicate if the
values in \code{x} represent single incursion weightings or arrival
probabilities respectively.}

\item{continued}{Logical to indicate whether incursions based on arrival
probabilities (when \code{type = "prob"}) continue to be generated during
spread simulations.}

\item{time_steps}{The time interval in simulation steps for generating
subsequent incursions when \code{type = "prob"} and
\code{continued = TRUE}.}

\item{...}{Additional parameters.}
}
\value{
An \code{Incursions} class object (list) containing functions for
accessing attributes and generating sampled incursions:
\describe{
\item{\code{get_type()}}{Get the incursion type.}
\item{\code{get_continued()}}{Get the continued indicator when type is
"prob".}
\item{\code{get_time_steps()}}{Get the continued incursion time steps
when applicable.}
\item{\code{generate()}}{Generates incursion locations via sampling.}
}
}
\description{
Builds a class to generate invasive species incursions, either by selecting
a single incursion location via stochastic sampling based on weightings for
each available spatial location, or by sampling potentially multiple
locations via incursion or arrival probabilities defined for each location,
which optionally may continue to generate incursions at specified time
intervals during a spread simulation.
}
