% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initializer.R
\name{Initializer}
\alias{Initializer}
\alias{Initializer.Raster}
\alias{Initializer.SpatRaster}
\alias{Initializer.default}
\alias{Initializer.Incursions}
\title{Initializer class builder}
\usage{
Initializer(
  x,
  region = NULL,
  population_model = NULL,
  class = character(),
  ...
)

\method{Initializer}{Raster}(x, ...)

\method{Initializer}{SpatRaster}(x, region = NULL, ...)

\method{Initializer}{default}(
  x,
  region = NULL,
  population_model = NULL,
  class = character(),
  ...
)

\method{Initializer}{Incursions}(
  x,
  region = NULL,
  population_model = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{x}{A \code{raster::Raster*}, \code{terra::SpatRaster}, or numeric
vector or array defining the initial population distribution (where layers
or columns optionally represent stages), or an \code{Incursions} or
inherited class object for generating initial, and optionally continued,
invasive species incursions. The initial population distribution may be
defined with actual populations sizes at appropriate locations (default
with no \code{size} attribute), or via the combination of a binary mask
indicating initial locations and an optional attached \code{size}
attribute indicating the size of the total initial population.}

\item{region}{A \code{Region} or inherited class object defining the spatial
locations included in the spread simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the spread simulations.}

\item{class}{Character class name for inherited classes. Default is empty.}

\item{...}{Additional parameters.}
}
\value{
An \code{Initializer} class object (list) containing a function to
appropriately initialize the population of each location:
\describe{
\item{\code{initialize()}}{Generates an appropriately represented
initial population for each simulated location via either predefined
values, or stochastic incursions.}
\item{\code{continued_incursions()}}{Produces a function for generating
continued incursions when configured to do so. The function produced
has the form: \code{function(tm, n)}, where \code{tm} is the
simulation time step and \code{n} is the array representing the
population at each location at that time. The function returns a
population array with merged incursions.}
}
}
\description{
Builds a class to initialize a spread simulation, given a specified spatial
region and population representation, via either a predefined initial
population distribution, or via stochastic generation of invasive species
incursions.
}
\references{
Bradhurst, R., Spring, D., Stanaway, M., Milner, J., & Kompas, T. (2021).
A generalised and scalable framework for modelling incursions,
surveillance and control of plant and environmental pests.
\emph{Environmental Modelling & Software}, 139, N.PAG.
\doi{10.1016/j.envsoft.2021.105004}

García Adeva, J. J., Botha, J. H., & Reynolds, M. (2012). A simulation
modelling approach to forecast establishment and spread of Bactrocera
fruit flies. \emph{Ecological Modelling}, 227, 93–108.
\doi{10.1016/j.ecolmodel.2011.11.026}
}
