% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{Kernels}
\alias{Kernels}
\title{Kernel functions generation class builder}
\usage{
Kernels(multiplier = NULL, ...)
}
\arguments{
\item{multiplier}{Optional numeric multiplier for scaling the (relative)
probabilities returned by the kernel functions.}

\item{...}{Additional parameters.}
}
\value{
A \code{Kernel} class object (list) containing functions for
generating distance or direction kernel functions for calculating the
(relative) probability of dispersal. The generated (returned) functions
of form \code{function(x)}, where \code{x} is a vector of distances (in m)
or directions (0-360 degrees), which return the (relative) probability for
each value in \code{x}:
\describe{
\item{\code{get_beta_function(alpha, beta, upper = 1)}}{Get a Beta
kernel function specified with distribution parameters \code{alpha}
and \code{beta}, as well as an optional parameter for
indicating the \code{upper} limit of \code{x} (default is 1).}
\item{\code{get_cauchy_function(scale)}}{Get a (half) Cauchy kernel
function specified with distribution parameter \code{scale} (for
scaling \code{x}).}
\item{\code{get_exp_function(mean)}}{Get a (negative) Exponential kernel
function specified with distribution parameter \code{mean}.}
\item{\code{get_gaussian_function(sd)}}{Get a (half) Gaussian (normal)
kernel function specified with distribution parameter \code{sd}
(standard deviation).}
\item{\code{get_lognormal_function(mean, sd)}}{Get a Lognormal kernel
function specified with distribution parameters \code{mean} and
\code{sd} (standard deviation).}
\item{\code{get_weibull_function(shape, scale)}}{Get a Weibull kernel
function specified with distribution parameters \code{shape} and
\code{scale} (for scaling \code{x}).}
\item{\code{get_lookup_function(table)}}{Get a table look-up kernel
function specified with parameter \code{table}, a two-column data
frame (or matrix) of \code{x} values (first column) and corresponding
(relative) probability values (second column). Look-up functionality
performs linear interpolation between discrete values.}
}
}
\description{
Builds a class for generating distance or direction functions (or kernels)
for calculating the (relative) probability of dispersal, used by
\code{Dispersal} models and inherited class objects. The kernel functions
include Beta, Cauchy, (negative) Exponential, Gaussian (normal), Lognormal,
and Weibull distributions. Also included is a function for defining the
probability of distances or directions from a table of values.
}
\references{
Carrasco, L. R., Harwood, T. D., Toepfer, S., MacLeod, A., Levay, N.,
Kiss, J., Baker, R. H. A., Mumford, J. D., & Knight, J. D. (2010).
Dispersal kernels of the invasive alien western corn rootworm and the
effectiveness of buffer zones in eradication programmes in Europe.
\emph{Annals of Applied Biology}, 156(1), 63–77.
\doi{10.1111/j.1744-7348.2009.00363.x}

Jongejans, E., Skarpaas, O., & Shea, K. (2008). Dispersal, demography and
spatial population models for conservation and control management.
\emph{Perspectives In Plant Ecology Evolution And Systematics}, 9(3–4),
153–170. \doi{10.1016/j.ppees.2007.09.005}

Shaw, M. W. (1995). Simulation of Population Expansion and Spatial Pattern
when Individual Dispersal Distributions do not Decline Exponentially with
Distance. \emph{Proceedings B: Biological Sciences}, 259(1356), 243–248.
\doi{10.1098/rspb.1995.0036}
}
