% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permeability.R
\name{Permeability}
\alias{Permeability}
\alias{Permeability.Raster}
\alias{Permeability.SpatRaster}
\title{Permeability class builder}
\usage{
Permeability(x, region, ...)

\method{Permeability}{Raster}(x, ...)

\method{Permeability}{SpatRaster}(x, region, ...)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial permeability of the spread simulation
region.}

\item{region}{A \code{Region} or inherited class object defining the spatial
locations included in the spread simulations.}

\item{...}{Additional parameters.}
}
\value{
A \code{Permeability} class object (list) containing functions for
accessing attributes:
\describe{
\item{\code{get_rast()}}{Get the permeability \code{terra::SpatRaster}
object.}
\item{\code{get_id()}}{Get the object numeric identifier.}
\item{\code{set_id(id)}}{Set the object numeric identifier.}
}
}
\description{
Builds a class to represent the permeability of cells/locations within the
spatial region of a spread simulation. The inverse of the permeability (0-1)
of cells is used to (linearly) scale the actual distance between adjacent
cells. For example, a permeability value of \code{0.5} results in an
effective distance twice that of the actual distance, a value of \code{0}
prevents spread to or through a cell, and a value of \code{1} does not
modify the effective distance.
}
