% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{Population}
\alias{Population}
\alias{Population.Raster}
\alias{Population.SpatRaster}
\alias{Population.Region}
\title{Population model class builder}
\usage{
Population(
  region,
  type = c("presence_only", "unstructured", "stage_structured"),
  growth = NULL,
  capacity = NULL,
  class = character(),
  ...
)

\method{Population}{Raster}(region, ...)

\method{Population}{SpatRaster}(region, ...)

\method{Population}{Region}(
  region,
  type = c("presence_only", "unstructured", "stage_structured"),
  growth = NULL,
  capacity = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{region}{A \code{raster::RasterLayer}, \code{terra::SpatRaster}, or
\code{Region} or inherited class object representing the spatial region
(template) for spread simulations.}

\item{type}{One of \code{"presence_only"} (default), \code{"unstructured"},
or \code{"stage_structured"} to indicate how populations are represented.}

\item{growth}{Numeric intrinsic growth rate or lambda (e.g. 1.2 for 20\%
growth per time step) when \code{type = "unstructured"}, or
age/stage-based transition rates (matrix) for each time step when
\code{type = "stage_structured"}. Default is \code{NULL} for when
\code{type = "presence_only"}.}

\item{capacity}{A data frame, \code{raster::RasterLayer}, or
\code{terra::SpatRaster} with values for the carrying capacity of the
invasive species at each location. Default is \code{NULL} for when
\code{type = "presence_only"}.}

\item{class}{Character class name for inherited classes. Default is empty.}

\item{...}{Additional parameters.}
}
\value{
A \code{Population} class object (list) containing function for
simulating growth:
\describe{
\item{\code{grow(x)}}{Perform growth or age/stage-based transition on
population \code{x} array (stages by locations), and return
transformed array.}
}
}
\description{
Builds a class for population representation and growth functionality for
the spread simulations.
}
