% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{Population}
\alias{Population}
\alias{Population.Region}
\title{Population model class builder}
\usage{
Population(
  region,
  type = c("presence_only", "unstructured", "stage_structured"),
  growth = NULL,
  capacity = NULL,
  establish_pr = NULL,
  incursion_mean = NULL,
  class = character(),
  ...
)

\method{Population}{Region}(
  region,
  type = c("presence_only", "unstructured", "stage_structured"),
  growth = NULL,
  capacity = NULL,
  establish_pr = NULL,
  incursion_mean = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for spread simulations.}

\item{type}{One of \code{"presence_only"} (default), \code{"unstructured"},
or \code{"stage_structured"} to indicate how populations are represented.}

\item{growth}{Numeric intrinsic growth rate or lambda (e.g. 1.2 for 20\%
growth per time step) when \code{type = "unstructured"}, or
age/stage-based transition rates (matrix) for each time step when
\code{type = "stage_structured"}. Default is \code{NULL} for when
\code{type = "presence_only"}.}

\item{capacity}{A vector of carrying capacity values of the invasive species
at each location specified by the \code{region}. Default is \code{NULL}
for when \code{type = "presence_only"} or growth is not capacity-limited.}

\item{establish_pr}{An optional vector of probability values (0-1) to
represent the likelihood of establishment at each location specified by
the \code{region}. This may be used to avoid transient/unsuccessful
incursion or migration arrivals from being presented in the simulation
results, and/or from subsequently contributing to spread in presence-only
models. Default is \code{NULL}.}

\item{incursion_mean}{Numeric mean population size for incursion locations.
The population size is sampled from the Poisson distribution for each
incursion location.}

\item{class}{Character class name for inherited classes. Default is
\code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{Population} class object (list) containing functions for
accessing attributes, conforming data to the appropriate population
representation, and simulating growth:
\describe{
\item{\code{get_type()}}{Get the population representation type.}
\item{\code{get_growth()}}{Get the growth rate or age/stage transition
matrix.}
\item{\code{get_stages()}}{Get the number of stages in a age/stage-based
representation.}
\item{\code{get_capacity(cells = NULL)}}{Get the carrying capacity as a
vector of values for each region location or optionally specified
region locations (indices).}
\item{\code{get_establish_pr(cells = NULL)}}{Get the establishment
probability as a vector of values for each region location or
optionally specified region locations (indices).}
\item{\code{make(initial, current, incursion)}}{Make a population vector
or array (rows:stages x columns:locations) via the defined population
representation using vectors or arrays of the \code{initial} or
\code{current} and \code{incursion} population at each region
location.}
\item{\code{grow(x)}}{Performs growth or age/stage-based transitions on
population \code{x} vector/array (stages by locations), and returns
the transformed vector/array.}
}
}
\description{
Builds a class for population representation and growth functionality for
spread simulations.
}
