% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresencePopulation.R
\name{PresencePopulation}
\alias{PresencePopulation}
\title{Presence-only population model class builder}
\usage{
PresencePopulation(region, spread_delay = NULL, ...)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for spread simulations.}

\item{spread_delay}{Number of simulation time steps before an incursion at
a newly established location can spread to other locations. This provides
an implicit mechanism for growth. Default \code{NULL} assumes no delay.}

\item{...}{Additional parameters.}
}
\value{
A \code{PresencePopulation} class object (list) containing functions
for accessing attributes and simulating (implicit) growth:
\describe{
\item{\code{get_type()}}{Get the population representation type.}
\item{\code{get_spread_delay()}}{Get the delay in spread ability.}
\item{\code{make(initial, current, incursion)}}{Make a population vector
using vectors of the \code{initial} or \code{current} and
\code{incursion} population at each region location.}
\item{\code{grow(x)}}{Performs implicit growth on the population
\code{x} vector via spread delay when present, and returns the
transformed vector.}
}
}
\description{
Builds a class for representing presence-only populations in spread
simulations. Growth is implicitly simulated via a delay in the ability of
newly established populations to spread.
}
