% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresencePopulation.R
\name{PresencePopulation}
\alias{PresencePopulation}
\title{Presence-only population model class builder}
\usage{
PresencePopulation(region, establish_pr = NULL, spread_delay = NULL, ...)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the spread simulations.}

\item{establish_pr}{An optional vector of probability values (0-1) to
represent the likelihood of establishment at each location specified by
the \code{region}. This may be used to avoid transient/unsuccessful
incursions or migrations from being presented in the simulation results,
and/or from subsequently contributing to spread in presence-only models.
Default is \code{NULL}.}

\item{spread_delay}{Number of simulation time steps before an incursion at
a newly established location can spread to other locations. This provides
an implicit mechanism for growth. Default \code{NULL} assumes no delay.}

\item{...}{Additional parameters.}
}
\value{
A \code{PresencePopulation} class object (list) containing functions
for accessing attributes and simulating (implicit) growth:
\describe{
\item{\code{get_region()}}{Get the region object.}
\item{\code{get_type()}}{Get the population representation type.}
\item{\code{get_establish_pr()}}{Get the establishment probability as a
vector of values for each location.}
\item{\code{get_spread_delay()}}{Get the delay in spread ability.}
\item{\code{make(initial, current, incursion)}}{Make a population vector
using vectors of the \code{initial} or \code{current} and
\code{incursion} population at each region location.}
\item{\code{grow(x)}}{Performs implicit growth on the population
\code{x} vector via spread delay when present, and returns the
transformed vector.}
}
}
\description{
Builds a class for representing presence-only populations in spread
simulations. Growth is implicitly simulated via a delay in the ability of
newly established populations to spread.
}
\references{
Bradhurst, R., Spring, D., Stanaway, M., Milner, J., & Kompas, T. (2021).
A generalised and scalable framework for modelling incursions,
surveillance and control of plant and environmental pests.
\emph{Environmental Modelling & Software}, 139, N.PAG.
\doi{10.1016/j.envsoft.2021.105004}

Jongejans, E., Skarpaas, O., & Shea, K. (2008). Dispersal, demography and
spatial population models for conservation and control management.
\emph{Perspectives In Plant Ecology Evolution And Systematics}, 9(3–4),
153–170. \doi{10.1016/j.ppees.2007.09.005}
}
