% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{Region}
\alias{Region}
\alias{Region.Raster}
\alias{Region.SpatRaster}
\title{Region class builder}
\usage{
Region(x, suitability = FALSE, ...)

\method{Region}{Raster}(x, ...)

\method{Region}{SpatRaster}(x, suitability = FALSE, ...)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial region (template) for spread simulations.}

\item{suitability}{Logical to indicate whether only suitable locations
(cells) with values greater than zero should be used in the simulations.}

\item{...}{Additional parameters.}
}
\value{
A \code{Region} class object (list) containing functions for
accessing attributes and checking compatibility of objects with the
region:
\describe{
\item{\code{get_template()}}{Get the spatial template (with zeros in
non-NA locations).}
\item{\code{get_indices()}}{Get cell indices of grid locations that are
included in the simulation.}
\item{\code{get_locations()}}{Get the number of locations (cells or
patches) that are included in the simulation.}
\item{\code{get_type()}}{Get the type of spatial representation: grid
(raster cells) or patch (network).}
\item{\code{is_compatible(y)}}{Check the compatibility of object
\code{y} with the region defined by \code{x}.}
}
}
\description{
Builds a class to represent a spatial region for a spread simulation,
defined by a raster layer with active (non-NA) and/or suitable (non-zero)
cells, or by a network of locations or patches.
}
