% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{Region}
\alias{Region}
\alias{Region.Raster}
\alias{Region.SpatRaster}
\title{Region class builder}
\usage{
Region(x, ...)

\method{Region}{Raster}(x, ...)

\method{Region}{SpatRaster}(x, ...)
}
\arguments{
\item{x}{A \code{raster::RasterLayer} or \code{terra::SpatRaster}
object representing the spatial region (template) for spread simulations.}

\item{...}{Additional parameters.}
}
\value{
A \code{Region} class object (list) containing functions for
accessing attributes, checking compatibility of objects with the
region, and to maintain and calculate spatial data for dispersal:
\describe{
\item{\code{get_template(empty = FALSE)}}{Get the spatial template with
either zeros in non-NA locations (default), or with no values when
\code{empty = TRUE}.}
\item{\code{get_indices()}}{Get cell indices of grid locations that are
included in the simulation.}
\item{\code{get_locations()}}{Get the number of locations (cells or
patches) that are included in the simulation.}
\item{\code{get_type()}}{Get the type of spatial representation: grid
(raster cells) or patch (network).}
\item{\code{is_compatible(y)}}{Check the compatibility of object
\code{y} with the region defined by \code{x}.}
\item{\code{is_included(indices)}}{Check if cell \code{indices} of grid
locations are included (non-NA cells) in the simulation (when type is
grid), and return a logical vector indicating the inclusion of each
index.}
\item{\code{two_tier()}}{Check if the region is configured for two-tier
dispersal (local plus aggregate cells) when type is grid.}
\item{\code{get_aggr()}}{Get a list of two-tier dispersal aggregation
components \code{factor}, \code{inner_radius}, \code{rast}
(\code{terra::SpatRaster}), \code{indices} (non-NA), \code{pts}
(\code{terra::SpatVector}), \code{cells} (list of region cells), and a
function \code{get_cells(indices)} that returns the region cells
(indices) within the specified aggregate cell (\code{indices}).}
\item{\code{set_aggr(aggr_factor, inner_radius)}}{Configure a two-tier
dispersal aggregation of the region via an aggregation factor and
an inner radius (in m) to define the boundary between local dispersal
at the original resolution and long-distance dispersal at an aggregate
resolution.}
\item{\code{configure_paths(directions = FALSE, max_distance = NULL,
      permeability = NULL)}}{Configure the inclusion of path directions from
occupied to reachable cells, an optional maximum distance (in m) for
the paths, and an optional \code{Permeability} class object containing
a spatial layer for calculating weighted paths, each of which are used
in dispersal calculations. Defaults infer non-inclusion.}
\item{\code{calculate_paths(cells)}}{Calculate the indices, distances
(in m), and directions (0-360 degrees) when required, of cells that
are reachable for each cell index in \code{cells}, including local and
aggregate cells when using two-tier configuration, and a graph
connecting these paths when configured with permeability layers.
The reachable cells are limited via a maximum distance (in m) and/or
permeability layers when either are configured.}
\item{\code{get_paths(cells, directions = FALSE, max_distance = NULL,
      perm_id = NULL)}}{Get a list of indices of, and distances (in m) and
(optional) directions (0-360 degrees) to, reachable cells for each
cell index in \code{cells}, including local (\code{$cell}) and
aggregate (\code{$aggr}) cells when configured, and optionally
(further) limited via a maximum distance (in m). Distances, and thus
reachable cells, are optionally modified via weighted paths specified
by a configured permeability object id.}
}
}
\description{
Builds a class to represent a spatial region for a spread simulation,
defined by a raster layer with active (non-NA) cells, or by a network of
locations or patches. When defined by a raster layer, the region can be
configured for two-tier dispersal, whereby long-distance dispersal is
calculated using a courser (aggregate) spatial resolution for destinations
outside of an inner radius, within which local dispersal is calculated using
the original spatial resolution of the region. Functions for calculating
distances and directions between (local and aggregate) cells are also
provided, as well as functionality for calculating weighted paths (graphs)
between cells (local and aggregate) via permeability layers. When used,
the inverse of the permeability (0-1) of cells is used to (linearly) scale
the actual distance between adjacent cells. The effective distance
between any two cells in the region is thus calculated via the shortest
weighted path between the cells.
}
