% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Removals.R
\name{Removals}
\alias{Removals}
\alias{Removals.Region}
\title{Removals class builder}
\usage{
Removals(
  region,
  population_model,
  removal_pr = 1,
  removal_pr_type = c("individual", "population"),
  remove_always = FALSE,
  detected_only = FALSE,
  removal_cost = NULL,
  radius = NULL,
  stages = NULL,
  schedule = NULL,
  ...
)

\method{Removals}{Region}(
  region,
  population_model,
  removal_pr = 1,
  removal_pr_type = c("individual", "population"),
  remove_always = FALSE,
  detected_only = FALSE,
  removal_cost = NULL,
  radius = NULL,
  stages = NULL,
  schedule = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the population spread model simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation for the population spread model
simulations.}

\item{removal_pr}{A single probability value (0-1), or vector of values for
each location specified by the \code{region}, to represent the likelihood
of removal at locations where the invasive species has been detected, as
indicated via a population attribute when present, else removal is applied
at all locations. Default is \code{1}, indicating all (detected)
occurrences are removed.}

\item{removal_pr_type}{Configures how the removal probability values are
specified. May be specified at either: the code{"individual"} level
(default), whereby removal probability values denote the probability of
removal for each (removable) invasive individual; the \code{"population"}
level, whereby removal probability values denote the probability of
removing all (removable) individuals within the (local) invasive
population.}

\item{remove_always}{A logical indication of whether or not removal is
always to applied to locations where invasive species are present (even
when they are not detected by explicit surveillance). Default is
\code{FALSE}, indicating that removal is dependent on detection when
surveillance actions are present. Set to \code{TRUE} for locations where
some removal is likely regardless of explicit management efforts (e.g.
via the general public or property owners).}

\item{detected_only}{A logical indication of whether or not removal is only
applied to detected individuals (e.g. via traps). Default is \code{FALSE},
indicating that removal is applied to all individuals at locations where
the invasive species has been detected (e.g. via treatment).}

\item{removal_cost}{Numeric vector of distributed removal costs (combined
resource and fixed costs) or a single cost value for each location where
removal is applied. For spatially-implicit area-based regions, cost
should be specified as cost per metres squared. Costs are accumulated for
each application of the removal at each (scheduled) simulation time step.
The cost unit may be added as an attribute
(\code{attr(removal_cost, "unit")}). Default is \code{NULL} when costs
are unavailable.}

\item{radius}{Optional radius (m) of the removal. Stochastic removal is
applied to all locations within the specified radius of each location
where the invasive species has been detected. Default is \code{NULL},
indicating that removal is only applied at detected locations (when
specified via a population attribute).}

\item{stages}{Numeric vector of population stages (indices) to which
removals are applied. Default is all stages (when set to \code{NULL}).}

\item{schedule}{Vector of discrete simulation time steps (t = 0, 1, 2, ...)
in which to apply removals. Default is all time steps (when set to
\code{NULL}).}

\item{...}{Additional parameters.}
}
\value{
A \code{Removals} class object (list) containing a function
for accessing attributes and applying simulated removals:
\describe{
\item{\code{get_type()}}{Get the type of action ("removal").}
\item{\code{get_id()}}{Get the actions numeric identifier.}
\item{\code{set_id(id)}}{Set the actions numeric identifier. Should be
an integer >= 1).}
\item{\code{get_label(include_id = TRUE)}}{Get the actions label used
in simulation results ("\if{html}{\out{<id>}}_removed" or "removed"). Set
\code{include_id} to include the action \code{id} as a label prefix
(default is \code{TRUE}).}
\item{\code{get_removal_pr_type()}}{Get the removal probability type.}
\item{\code{get_stages()}}{Get the population stages to which removals
are applied.}
\item{\code{get_schedule()}}{Get the scheduled simulation time steps in
which removals are applied.}
\item{\code{include_cost()}}{Logical indication of a cost parameter
having a value.}
\item{\code{get_cost_label(include_id = TRUE)}}{Get the removal
cost label used in simulation results ("\if{html}{\out{<id>}}_removal_cost" or
"removal_cost"). Set \code{include_id} to include the action \code{id}
as a label prefix (default is \code{TRUE}).}
\item{\code{get_cost_unit()}}{Get the unit of removal cost.}
\item{\code{get_attributes(n)}}{Get attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return a list of the attached attributes.}
\item{\code{clear_attributes(n)}}{Clear attached attributes associated
with this action from a simulated population vector or matrix
\code{n}, and return \code{n} without the attached attributes.}
\item{\code{apply(n, tm)}}{Apply removals to a simulated population
vector or matrix \code{n}, potentially with attached attributes
relating to previously applied actions, providing the time step
\code{tm} is in the \code{schedule}, and return the resulting
population \code{n} along with attached attributes relating to the
newly applied removals.}
}
}
\description{
Builds a class for simulating the application of management removals or
eradication of an invasive species.
}
