% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\name{Results}
\alias{Results}
\alias{Results.Region}
\title{Results class builder}
\usage{
Results(
  region,
  population_model,
  time_steps = 1,
  step_duration = 1,
  step_units = "years",
  collation_steps = 1,
  replicates = 1,
  combine_stages = NULL,
  ...
)

\method{Results}{Region}(
  region,
  population_model,
  time_steps = 1,
  step_duration = 1,
  step_units = "years",
  collation_steps = 1,
  replicates = 1,
  combine_stages = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the spread simulations.}

\item{population_model}{A \code{Population} or inherited class object
defining the population representation and growth functionality for the
spread simulations.}

\item{time_steps}{The number of discrete time steps to simulate. Default is
1.}

\item{step_duration}{The duration of the simulation time steps in units
specified by \code{step_units}. Default is 1.}

\item{step_units}{The units for the simulation step duration
(\code{step_duration}) as a character string. Default is "years".}

\item{collation_steps}{The interval in time steps for collating results.
Default is 1, that is, results are collated at every time step.}

\item{replicates}{The number of replicate or repeated simulations to be run.
Default is 1. Note that replicate simulations results are collated as
summary statistics across simulations.}

\item{combine_stages}{Optionally combine (sum) specified stages (a vector of
stage indices) of stage-based population results. The default
(\code{NULL}) maintains results for each stage.}

\item{...}{Additional parameters.}
}
\value{
A \code{Results} class object (list) containing functions for
calculating and collating results, as well as accessing lists of results
and the simulation parameters used to produce them:
\describe{
\item{\code{collate(r, tm, n)}}{Collate the results at simulation
replicate \code{r} and time step \code{tm} using the current vector or
array \code{n} representing the population at each location.}
\item{\code{finalize()}}{Finalize the results collation (summary
calculations).}
\item{\code{as_list()}}{Return the results as a list (collated, total,
area).}
\item{\code{get_params()}}{Get the simulation parameters used.}
\item{\code{save_rasters(...)}}{Save the collated results as raster TIF
files when the region is grid-based. \code{Terra} raster write options
may be passed to the function.}
\item{\code{save_csv()}}{Save the collated results as comma-separated
values (CSV) files when the region is patch-based. Also saves the
population totals and area occupied to CSV files for both grid and
patch-based region types.}
\item{\code{save_plots()}}{Save plots of the population (staged) totals
and the area occupied as PNG files.}
}
}
\description{
Builds a class for encapsulating, calculating, and collating spread
simulation results, including the population at each location at each
collation time step, the total population size, and the area occupied at
each time step. When simulations are replicated, summary results (means and
standard deviations) are produced.
}
