% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StagedPopulation.R
\name{StagedPopulation}
\alias{StagedPopulation}
\title{Stage or age-based population model class builder}
\usage{
StagedPopulation(
  region,
  growth,
  capacity = NULL,
  capacity_stages = NULL,
  establish_pr = NULL,
  incursion_values = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the spread simulations.}

\item{growth}{A stage or age-based matrix containing transition (fecundity
and survival) rates for each time step.}

\item{capacity}{A vector of carrying capacity values of the invasive species
at each location specified by the \code{region}. Default is \code{NULL}
for when growth is not capacity-limited.}

\item{capacity_stages}{A vector of stage or age indices (as per the
rows/columns of \code{growth} matrix) indicating which life-stages/ages
are applicable for capacity-limited growth (and survival). Default is
\code{NULL} for when growth is not capacity-limited, or when all
stages/ages are applicable for capacity-limited growth.}

\item{establish_pr}{An optional vector of probability values (0-1) to
represent the likelihood of establishment at each location specified by
the \code{region}. This may be used to avoid transient/unsuccessful
incursions or migrations from being presented in the simulation results.
Default is \code{NULL}.}

\item{incursion_values}{Defines how incursion locations are populated.
A matrix of fixed values with a row for each region location, or a single
row applicable for all locations, and with columns for each stage/age.
Alternatively, a list of matrices (as before), labelled \code{min} and
\code{max}, be used to specifying ranges of values for uniform random
sampling of values.}

\item{...}{Additional parameters.}
}
\value{
A \code{StagedPopulation} class object (list) containing functions
for accessing attributes and simulating growth:
\describe{
\item{\code{get_type()}}{Get the population representation type.}
\item{\code{get_growth()}}{Get the growth stage/age matrix.}
\item{\code{get_capacity()}}{Get the carrying capacity as a vector of
values for each location.}
\item{\code{get_capacity_stages()}}{Get the stage/age (indices) that are
applicable for capacity-limited growth.}
\item{\code{get_establish_pr()}}{Get the establishment probability as a
vector of values for each location.}
\item{\code{make(initial, current, incursion)}}{Make a population matrix
(with a row per location and a column per stage/age) via using vectors
or matrices of the \code{initial} or \code{current} and
\code{incursion} population at each region location.}
\item{\code{grow(x)}}{Performs logistic (capacity-limited) growth on the
population \code{x} matrix (having a row per location and a column per
stage/age), and returns the transformed matrix.}
}
}
\description{
Builds a class for representing stage or age-based populations in spread
simulations. Simulates logistic (capacity-limited) growth.
}
