% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnstructPopulation.R
\name{UnstructPopulation}
\alias{UnstructPopulation}
\title{Unstructured population model class builder}
\usage{
UnstructPopulation(
  region,
  growth = 1,
  capacity = NULL,
  establish_pr = NULL,
  incursion_mean = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the spread simulations.}

\item{growth}{Numeric intrinsic growth rate or lambda (e.g. 1.2 for 20\%
growth per time step). Default is \code{1.0} for no increase.}

\item{capacity}{A vector of carrying capacity values of the invasive species
at each location specified by the \code{region}. Default is \code{NULL}
for when growth is not capacity-limited.}

\item{establish_pr}{An optional vector of probability values (0-1) to
represent the likelihood of establishment at each location specified by
the \code{region}. This may be used to avoid transient/unsuccessful
incursions or migrations from being presented in the simulation results.
Default is \code{NULL}.}

\item{incursion_mean}{Numeric mean population size for incursion locations.
The population size is sampled from the Poisson distribution for each
incursion location.}

\item{...}{Additional parameters.}
}
\value{
A \code{UnstructPopulation} class object (list) containing functions
for accessing attributes and simulating growth:
\describe{
\item{\code{get_type()}}{Get the population representation type.}
\item{\code{get_growth()}}{Get the growth rate.}
\item{\code{get_capacity()}}{Get the carrying capacity as a vector of
values for each location.}
\item{\code{get_establish_pr()}}{Get the establishment probability as a
vector of values for each location.}
\item{\code{make(initial, current, incursion)}}{Make a population vector
via using vectors of the \code{initial} or \code{current} and
\code{incursion} population at each region location.}
\item{\code{grow(x)}}{Performs logistic (capacity-limited) growth on the
population \code{x} vector, and returns the transformed vector.}
}
}
\description{
Builds a class for representing unstructured populations in spread
simulations. Simulates logistic (capacity-limited) growth.
}
